### 2017-06-04 #278 @whedon
[PRE REVIEW]: opty: Software For Trajectory Optimization and Parameter Identification Using Direct Collocation
**Submitting author:** @moorepants (<a href="http://orcid.org/0000-0002-8698-6143">Jason K. Moore</a>)
**Repository:** <a href="https://github.com/csu-hmc/opty" target ="_blank">https://github.com/csu-hmc/opty</a>
**Version:** v0.2.0
