### 2017-06-06 ea7e054
Added Ton as a coauthor in the JOSS file.
