### 2017-06-20 #278 @Kevin-Mattheus-Moerman
@jakevdp Do you want to review this? If yes, great I can remain editor. If not, could you be editor and/or help find a reviewer?
