### 2017-06-20 #278 @moorepants
Here are some potential reviewer sources:

Main contributors to:
- Casadi: https://github.com/casadi/casadi/graphs/contributors
- Opensim: https://github.com/opensim-org/opensim-core/graphs/contributors
- BTK: https://github.com/Biomechanical-ToolKit/BTKCore/graphs/contributors
- Simbody: https://github.com/simbody/simbody/graphs/contributors
- Drake: https://github.com/RobotLocomotion/drake/graphs/contributors
- psopt: https://github.com/PSOPT/psopt/graphs/contributors
- acado: https://github.com/acado/acado/graphs/contributors
