### 2017-06-20 #278 @moorepants
I'm pinging my co-author @tvdbogert so that he is subscribed to this thread.
