### 2017-06-20 #300 @whedon
[REVIEW]: opty: Software For Trajectory Optimization and Parameter Estimation Using Direct Collocation
**Submitting author:** @moorepants (<a href="http://orcid.org/0000-0002-8698-6143">Jason K. Moore</a>)
**Repository:** <a href="https://github.com/csu-hmc/opty" target ="_blank">https://github.com/csu-hmc/opty</a>
**Version:** v0.2.0
