### 2017-06-20 #300 @moorepants
It is worth nothing that Chris is a co-author of mine (at least in the last few years) and we are good friends. I'm not sure of JOSS rules, but in other journals this would likely be a conflict of interest. FYI.
