### 2017-06-21 ca686bb
Removed ancient Dockerfile.
