### 2017-07-13 #300 @moorepants
Just checking in on this. It's been well over a month since I submitted.
