### 2017-07-17 #300 @Kevin-Mattheus-Moerman
@demotu @siboles @stavness @melund @jslee02 I am looking for an additional reviewer for [this submission](https://github.com/csu-hmc/opty). Would you be interested in reviewing this for [JOSS](http://joss.theoj.org/)? The review guidelines are [here](http://joss.theoj.org/about#reviewer_guidelines).
