### 2017-07-17 #300 @stavness
Hi Kevin,

I'd be happy to help -- I like the idea of JOSS. I'm on vacation for the
next 3 weeks though, so I wouldn't be able to get it done until ~mid
August. Let me know if that's ok for your timeline.

Ian

On Sun, Jul 16, 2017 at 7:02 PM, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @demotu <https://github.com/demotu> @siboles <https://github.com/siboles>
> @stavness <https://github.com/stavness> @melund
> <https://github.com/melund> @jslee02 <https://github.com/jslee02> I am
> looking for an additional reviewer for this submission
> <https://github.com/csu-hmc/opty>. Would you be interested in reviewing
> this for JOSS <http://joss.theoj.org/>? The review guidelines are here
> <http://joss.theoj.org/about#reviewer_guidelines>.
>
