### 2017-07-17 #300 @Kevin-Mattheus-Moerman
Hey @stavness. Mid-August is fine. By then we'll hopefully have some comments from the first reviewer also. Enjoy your vacation.
