### 2017-07-17 #300 @jslee02
I'm currently in the middle of moving to a different city so my timeline would be similar to @stavness for settling down at the new place. If it's fine and still needs an additional reviewer, then I'd be happy to participate. Just let me know what you think.
