### 2017-07-17 #300 @melund
Hi @Kevin-Mattheus-Moerman. Sorry for slow reply. I am also on holiday at the moment. I see you got your second reviewer, but I don't mind helping out some other time if it is biomechanics related. Funny, I had just watched Kyles presentation on Scipy when I got this notification. JOSS is a much needed initiative.
