### 2017-07-17 #300 @Kevin-Mattheus-Moerman
@jslee02 @melund Thanks for getting back to me. It looks like most reviewers have vacation etc. so the mid-august time-line is fine. If this submission is of interest to  you, and you have time, I would greatly appreciate if you could also help to review this submission. Our formal system handles 1 "official reviewer". But you can review by leaving comments here (i.e. based on the tick-marks at the top of this issue and http://joss.theoj.org/about#reviewer_guidelines).
