### 2017-07-17 #300 @demotu
Hi @Kevin-Mattheus-Moerman <https://github.com/kevin-mattheus-moerman>

Here is @demotu <https://github.com/demotu>. I also can review the
submission. This work interests to me so even if there are enough potential
reviewers, I will look at it.

-
Marcos Duarte
http://demotu.org/

On Sun, Jul 16, 2017 at 10:02 PM, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @demotu <https://github.com/demotu> @siboles <https://github.com/siboles>
> @stavness <https://github.com/stavness> @melund
> <https://github.com/melund> @jslee02 <https://github.com/jslee02> I am
> looking for an additional reviewer for this submission
> <https://github.com/csu-hmc/opty>. Would you be interested in reviewing
> this for JOSS <http://joss.theoj.org/>? The review guidelines are here
> <http://joss.theoj.org/about#reviewer_guidelines>.
>
