### 2017-07-31 #300 @cybanical
@moorepants Is this only working on a 32-bit install? I just did a fresh anaconda py3.6 install and got this error:
```
PackageNotFoundError: Package missing in current win-64 channels:
  - opty

```
Also, windows anaconda Anaconda is not aware of ipopt, cyipopt.
It looks like i need to install ipopt separately. This requires cygwin or mingw.

Perhaps there could be a bit more clarity/handholding on windows based installation?
