### 2017-08-05 #300 @moorepants
@cybanical Thanks, but i do not support Windows at the moment. You can manually install ipopt and then cyipopt and it should work but I do not have the resources to get it working on windows. I suggest using a linux VM or docker image and using the install instructions in the readme.
