### 2017-08-06 #300 @cybanical
@moorepants In that case it might make sense to explicitly state that windows is not supported, rather than give partial instructions for a windows installation - it'd save some headache.
