### 2017-08-06 #300 @Kevin-Mattheus-Moerman
@moorepants can you more clearly state if Windows is actively supported?
@cybanical will you still be able to review this submission? If needed I can provide you with a Linux environment for testing this software.
