### 2017-08-13 #300 @moorepants
I'm pretty sure that the software will work fine on Windows but have never tested it myself. The user would need to manually install all of the dependencies as per their Windows specific instructions. IPOPT, NumPy, SciPy, etc are not necessarily trivial to compile on Windows but one likely can find binaries. I'm happy to do bug fixes wrt to Windows but do not plan on developing installation instructions for all of the dependencies in the opty documentation, as this can be found in the respective packages' docs. I'm happy to make this clearer in the README.
