### 2017-08-16 #300 @demotu
I am reviewing this submission; it will be done by the weekend.

@demotu <https://github.com/demotu>

-
Marcos Duarte
http://demotu.org/

On Sat, Aug 12, 2017 at 11:05 PM, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @stavness <https://github.com/stavness> @jslee02
> <https://github.com/jslee02> @melund <https://github.com/melund> @demotu
> <https://github.com/demotu> we are getting close to mid-august 😄 I hope
> those who had a vacation had a good one. If you are still able, would you
> be willing to help review this submission? Thanks!
>
