### 2017-08-21 #300 @stavness
I installed the software and ran the examples. Everything went smoothly on
ubuntu. Will read the paper this week.

On Sun, Aug 20, 2017 at 8:17 PM Kevin Mattheus Moerman <
notifications@github.com> wrote:

> Thanks @demotu <https://github.com/demotu> you can leave review comments
> here. Alternatively you can open issues on the repository for this
> submission <https://github.com/csu-hmc/opty> and refer to the issues here.
>
