### 2017-08-21 #300 @Kevin-Mattheus-Moerman
@stavness Great, thanks Ian! Let me know if you have any comments on the other aspects, such as documentation (see tickmarks at the top of this issue).
