### 2017-08-21 #300 @demotu
I tested optpy <https://github.com/csu-hmc/opty> in a Linux machine (Ubuntu
64 bits with a conda installation of Python 3.6).
The examples ran fine, with the exception of the inverted_pendulum.py; I
opened an issue on its repo.
The software and the paper comply with all the JOSS guidelines (
http://joss.theoj.org/about#reviewer_guidelines).
The paper (a summary with a few lines) is in markdown format, not in PDF as
described in the JOSS guidelines, but this is not a problem.
My congratulations to the authors for the development of optpy.

I recommend optpy <https://github.com/csu-hmc/opty> for publication in JOSS.

regards,

-
Marcos Duarte
http://demotu.org/

On Sun, Aug 20, 2017 at 11:17 PM, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> Thanks @demotu <https://github.com/demotu> you can leave review comments
> here. Alternatively you can open issues on the repository for this
> submission <https://github.com/csu-hmc/opty> and refer to the issues here.
>
