### 2017-08-28 #300 @stavness
Sorry is there also a paper, or just the git repo?

On Sun, Aug 20, 2017 at 8:47 PM, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @stavness <https://github.com/stavness> Great, thanks Ian! Let me know if
> you have any comments on the other aspects, such as documentation (see
> tickmarks at the top of this issue).
>
