### 2017-08-29 #300 @stavness
Thanks for the clarification Kevin. Please find my review below.

The paper is missing a number of required components according to the JOSS
guidelines:

* A summary describing the high-level functionality and purpose of the
software for a diverse, non-specialist audience
— current summary is written in highly technical language, I suggest a lay
introduction followed by a more detailed technical description of the
software that includes equations to clarify what types of problems are
applicable for the software
— [optional] a full theory guide would be very beneficial for introducing
the basic concepts of direct colocation. Although this type of information
can be found in textbooks / papers (e.g. the Betts paper cited), but it can
be very helpful to include it within the software paper, as the software
authors typically have their own interpretation of the theory, and the
equations can be better matched to the software implementation. For
example, both FEBio and Simbody have excellent theory guides, which are
arguably one of the main contributions of those works.

* A clear statement of need that illustrates the purpose of the software
- Motivation should be expanded, why is this software useful, what targeted
applications?

* A list of key references including a link to the software archive
- References are not rendered in the paper.md file
- No link to software archive
- Other than Betts2010, the other citations appear to be just for
dependencies. The paper should include citations to other software for
direct colocation, and a comparison of feature sets, etc.

* Mentions (if applicable) of any ongoing research projects using the
software or recent scholarly publications enabled by it
- Needs to be added

Software license:
- acceptable
-
Documentation:
- acceptable

API documentation
- acceptable

Tests
- acceptable — uses Travis-CI

Community guidelines
- Missing, should include:
There should be clear guidelines for third-parties wishing to:
Contribute to the software
Report issues or problems with the software
Seek support
Examples

Functionality
- Installation and examples worked

On Tue, Aug 29, 2017 at 7:42 AM, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @stavness <https://github.com/stavness> There is a paper.md file
> <https://github.com/csu-hmc/opty/blob/master/paper.md> which will be
> rendered into a PDF later on. Since the software documentation/API etc
> contains most of the technical descriptions the papers in JOSS are very
> short (see also our paper guidelines
> <http://joss.theoj.org/about#author_guidelines>).
>
