### 2017-08-29 #300 @moorepants
@stavness Thank you for the helpful review. We will work on addressing these items. They all seem reasonable, except for less-than-clear scope of a theory guide. I agree that the Simbody (have only read that one) guide is a top notch contribution, but that kind of guide seems to be an outlier. I've only seen that kind of detail in the most exceptional packages and it would take a large amount of time to create something similar. I'd rather cite textbooks/articles for this information and simply explain our implementation wrt to those. For example, I think the IPOPT software has a minimal explanation of the theory and cites the relevant literature, which I'd be more keen to replicate.
