### 2017-08-30 #300 @Kevin-Mattheus-Moerman
@stavness I've spoken with @cybanical who was the primary reviewer here. He does not currently have time to continue this review so I kindly ask if you could take over. Essentially you should now be able to formally tick the boxes at the top of this review issue.
