### 2017-08-30 #300 @stavness
ok I've accepted the invitation

On Wed, Aug 30, 2017 at 10:24 AM, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @stavness <https://github.com/stavness> I've spoken with @cybanical
> <https://github.com/cybanical> who was the primary reviewer here. He does
> not currently have time to continue this review so I kindly ask if you
> could take over. Essentially you should now be able to formally tick the
> boxes at the top of this review issue.
>
