### 2017-08-31 #300 @jslee02
Sorry for the late response. Below is my review based on the template. There is only one missing component according to the JOSS guidelines.

- [x] **Repository**: Is the source code for this software available at the repository url?
- [x] **License**: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
  - **comment**: The license is not being detected by Github. It would be nice to adjust `LICENSE.txt` so as to be displayed on the repository overview. (see: https://github.com/blog/2252-license-now-displayed-on-repository-overview)
- [x] **Version**: Does the release version given match the GitHub release (v0.2.0)?
- [x] **Authorship**: Has the submitting author (@moorepants) made major contributions to the software?

- [x] **Installation**: Does installation proceed as outlined in the documentation?
- [x] **Functionality**: Have the functional claims of the software been confirmed?
- [x] **Performance**: Have any performance claims of the software been confirmed?

- [x] **Installation instructions**: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
- [x] **Example usage**: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
- [x] **Functionality documentation**: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
- [x] **Automated tests**: Are there automated tests or manual steps described so that the function of the software can be verified?
- [ ] **Community guidelines**: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
  - **comment**: Missing. It would be great to have one. There is a useful instruction on the de facto standard way to add a contribution guideline: https://help.github.com/articles/setting-guidelines-for-repository-contributors/.

- [x] **Authors**: Does the paper.md file include a list of authors with their affiliations?
- [x] **References**: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
  - **comment**: Nitpick: According to the [author guidelines of JOSS](http://joss.theoj.org/about#author_guidelines), it seems `paper.bib` is the recommended name for the file. Is there a reason that `joss.bib` was used instead?

- The link at the top of the repository is http://hmc.csuohio.edu that I believe that it refers to the associated lab's website rather than the software's website. At first, I thought this is because the software website is missing but later found the link to it in `README.md` that is http://opty.readthedocs.io/. It would be great to put the link to the software website at the top of the repository, which is meant to be used so, and move the link to the lab website to somewhere else.
- All the examples worked except `examples/inverted_pendulum` printing:
```
$ python pendulum.py
  File "pendulum.py", line 48, in <module>
    import simulate
  File "/Users/jeongseoklee/dev/opty/examples/inverted_pendulum/simulate.py", line 10, in <module>
    from pydy.codegen.code import generate_ode_function
ModuleNotFoundError: No module named 'pydy.codegen.code'
```
