### 2017-10-16 #300 @Kevin-Mattheus-Moerman
@moorepants were you able to address @stavness comments?

@stavness Are you able to tick some of the boxes at the top of this issue?

@jslee02 thanks for your review! @moorepants can you address these comments as well?
