### 2017-10-16 #300 @moorepants
I am planning to address the comments. I have been bogged down with hosting a conference and then prepping two new courses. Things should calm down soon. Thanks for the reminder.
