### 2017-10-19 csu-hmc/opty#58 @moorepants
State that WIndows is not explicitly supported
From https://github.com/openjournals/joss-reviews/issues/300#issuecomment-320479419

> @moorepants In that case it might make sense to explicitly state that windows is not supported, rather than give partial instructions for a windows installation - it'd save some headache.
