### 2017-10-19 csu-hmc/opty#59 @moorepants
Improve the high level summary
From: https://github.com/openjournals/joss-reviews/issues/300#issuecomment-325696614

> A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience — current summary is written in highly technical language, I suggest a lay introduction followed by a more detailed technical description of the software that includes equations to clarify what types of problems are applicable for the software
