### 2017-10-19 csu-hmc/opty#60 @moorepants
Improve the theory explanation in the docs
From: https://github.com/openjournals/joss-reviews/issues/300#issuecomment-325696614
> — [optional] a full theory guide would be very beneficial for introducing the basic concepts of direct colocation. Although this type of information can be found in textbooks / papers (e.g. the Betts paper cited), but it can be very helpful to include it within the software paper, as the software authors typically have their own interpretation of the theory, and the equations can be better matched to the software implementation. For example, both FEBio and Simbody have excellent theory guides, which are arguably one of the main contributions of those works.
