### 2017-10-19 csu-hmc/opty#61 @moorepants
Improve statement of need
From: https://github.com/openjournals/joss-reviews/issues/300#issuecomment-325696614
> A clear statement of need that illustrates the purpose of the software - Motivation should be expanded, why is this software useful, what targeted applications?
