### 2017-10-19 csu-hmc/opty#62 @moorepants
Add list of references
From: https://github.com/openjournals/joss-reviews/issues/300#issuecomment-325696614
> A list of key references including a link to the software archive - References are not rendered in the paper.md file - No link to software archive - Other than Betts2010, the other citations appear to be just for dependencies. The paper should include citations to other software for direct colocation, and a comparison of feature sets, etc.

- [x] finish up the table of comparisons
