### 2017-10-19 csu-hmc/opty#63 @moorepants
Mention related research projects
From: https://github.com/openjournals/joss-reviews/issues/300#issuecomment-325696614
> Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it - Needs to be added
