### 2017-10-19 csu-hmc/opty#64 @moorepants
Add contributing guidelines
From: https://github.com/openjournals/joss-reviews/issues/300#issuecomment-325696614
> Community guidelines - Missing, should include: There should be clear guidelines for third-parties wishing to: Contribute to the software Report issues or problems with the software Seek support Examples

From: https://github.com/openjournals/joss-reviews/issues/300#issuecomment-326260226
> comment: Missing. It would be great to have one. There is a useful instruction on the de facto standard way to add a contribution guideline: https://help.github.com/articles/setting-guidelines-for-repository-contributors/.
