### 2017-10-19 csu-hmc/opty#65 @moorepants
Make license detected by Github
From: https://github.com/openjournals/joss-reviews/issues/300#issuecomment-326260226

> comment: The license is not being detected by Github. It would be nice to adjust LICENSE.txt so as to be displayed on the repository overview. (see: https://github.com/blog/2252-license-now-displayed-on-repository-overview)
