### 2017-10-19 csu-hmc/opty#66 @moorepants
Rename joss.bib
From: https://github.com/openjournals/joss-reviews/issues/300#issuecomment-326260226
> comment: Nitpick: According to the author guidelines of JOSS, it seems paper.bib is the recommended name for the file. Is there a reason that joss.bib was used instead?
