### 2017-10-19 csu-hmc/opty#67 @moorepants
Official website
From: https://github.com/openjournals/joss-reviews/issues/300#issuecomment-326260226
> The link at the top of the repository is http://hmc.csuohio.edu that I believe that it refers to the associated lab's website rather than the software's website. At first, I thought this is because the software website is missing but later found the link to it in README.md that is http://opty.readthedocs.io/. It would be great to put the link to the software website at the top of the repository, which is meant to be used so, and move the link to the lab website to somewhere else.
