### 2017-10-19 csu-hmc/opty#68 @moorepants
Inverted pendulum example no longer works
From: https://github.com/openjournals/joss-reviews/issues/300#issuecomment-326260226
> All the examples worked on macOS Sierra except examples/inverted_pendulum printing:
