### 2017-10-19 ee82e6f
Added note to the README stating that only Unix is officially supported.
Fixes issue #58 for the JOSS submission.
