### 2017-10-19 csu-hmc/opty#65 @moorepants
This is the software that does the detection: https://github.com/benbalter/licensee

It isn't clear to me why this wouldn't be detected. Maybe it doesn't like the `.txt` extension?
