### 2017-10-19 8085a02
Renamed LICENSE.txt to LICENSE
Fixes issue #65 for the JOSS review.
