### 2017-10-19 27142f6
Added a CONTRIBUTING file.
Fixes issue #64 for the JOSS review.
