### 2017-10-19 csu-hmc/opty#66 @moorepants
I probably named it joss.bib because it annoys me that JOSS requires you to dump generic named files in the root of the repo, see: https://github.com/openjournals/joss/issues/255.
