### 2017-10-19 90c0f2b
Renamed joss.bib to paper.bib.
Fixes issue #66 for the JOSS review.
