### 2017-10-19 779e2b6
Added HMC funding to readme.
