### 2017-10-19 f2cce00
Mentioned a related research project at CSU.
Fixes issue 63 for the JOSS review.
