+++ paper.bib
@@ -1,3 +1,4 @@
+% Encoding: UTF-8
 
 @article{Wachter2006,
 	title = {On the implementation of an interior-point filter line-search algorithm for large-scale nonlinear programming},
@@ -69,3 +70,13 @@ DOI: 10.1137/1.9780898718577},
 	author = {{cyipopt Developers}},
 	year = {2017}
 }
+
+@InProceedings{Wang2017,
+  author    = {Huawei Wang and Anne Koelewijn and Antonie van den Bogert},
+  title     = {RAMP PERTURBATION TESTS ARE TOO SIMPLE TO IDENTIFY A REALISTIC CONTROLLER IN HUMAN STANDING BALANCE},
+  booktitle = {Dynamic Walking},
+  year      = {2017},
+  url       = {http://ruina.tam.cornell.edu/dynwalk17abstracts/258-Anne%20Koelewijn%20-%20HuaweiAbstract.pdf},
+}
+
+@Comment{jabref-meta: databaseType:bibtex;}
