### 2017-10-19 csu-hmc/opty#62 @moorepants
- Casadi
- Gpops
