### 2017-10-19 csu-hmc/opty#65 @moorepants
I removed the `.txt` extension. Not sure what else to do. Closing for now.
