### 2017-10-19 csu-hmc/opty#62 @moorepants
@tvdbogert If you have any to add to this list here, feel free to. I'm not sure I'm that equipped to do a feature comparison, as I haven't used the other software. But I'll give it a shot.
