### 2017-10-19 csu-hmc/opty#62 @moorepants
@tvdbogert Your paragraph looks good. Let's add in the references.
