### 2017-10-20 11b7e87
Improved paper.md and added comparison table.
