+++ paper.md
@@ -25,27 +25,36 @@ bibliography: paper.bib
 
 # Summary
 
-opty operates on symbolic descriptions of ordinary differential equations to
-form the constraints needed to solve trajectory optimization and parameter
-identification problems using the direct collocation method [@Betts2010]. The
-direct collocation problem is then solved using non-linear programming through
-cyipopt [@Cyipopt2017] and IPOPT [@Watcher2006]. The software allows the user to
-describe the dynamical system of interest at a high level in symbolic form
-without needing to concern themselves with the numerical computation details.
-To do this, SymPy [@Meurer2017] and Cython [@Behnel2011] are utilized for code
-generation and just-in-time compilation to obtain wrap optimized C functions
-for use in Python. The target audience are engineers and scientists interested
-in solving nonlinear optimal control and parameter identification problems with
-minimal overhead.
+opty is a tool for describing and solving trajectory optimization and parameter
+identification problems based on symbolic descriptions of ordinary differential
+equations that describe a dynamical system. The motivation for its development
+resides in the need to solve optimal control problems of biomechanical systems.
+The target audience is engineers and scientists interested in solving nonlinear
+optimal control and parameter identification problems with minimal
+computational overhead.
+
+A user of opty is responsible for specifying the system's dynamics (the ODEs),
+the cost or fitness function, bounds on the solution, and the initial guess for
+the solution. opty uses this problem specification to derive the constraints
+needed to solve the optimization problem using the direct collocation method
+[@Betts2010]. This method maps the problem to a non-linear programming problem
+and the result is then solved numerically with an interior point optimizer,
+IPOPT [@Watcher2006] which is wrapped by cyipopt [@Cyipopt2017] for use in
+Python. The software allows the user to describe the dynamical system of
+interest at a high level in symbolic form without needing to concern themselves
+with the numerical computation details. This is made possible by utilizing
+SymPy [@Meurer2017] and Cython [@Behnel2011] for code generation and
+just-in-time compilation to obtain wrap optimized C functions that are
+accessible in Python.
 
 Direct collocation methods have been especially successful in the field of
 human movement [@Ackermann2010, @vandenBogert2012] because those systems are
 highly nonlinear, dynamically stiff, and unstable with open loop control.
-Typically, closed-source tools were used for multibody dynamics (SD/Fast,
+Typically, closed-source tools were used for multibody dynamics (e.g. SD/Fast,
 Autolev), for optimization (SNOPT), and for the programming environment
 (Matlab). Recently, promising work has been done with the Opensim/Simbody
 dynamics engine [@LeeUmberger2016, @LinPandy2017], but this requires that
-Jacobian matrices are approximated by finite derivatives. In contrast, opty
+Jacobian matrices are approximated by finite differences. In contrast, opty
 provides symbolic differentiation which makes the code faster and prevents poor
 convergence when the severe nonlinearity causes finite differences to be
 inaccurate. Furthermore, opty allows the system to be formulated using an
@@ -63,7 +72,36 @@ however, the low order discretizations provide more robust convergence to the
 globally optimal trajectory in a nonlinear system when a good initial guess is
 not available [@Zarei2016].
 
-Feature comparison (this should be in a table if we do it)
+There are existing software packages that have similarities to opty. Below, is
+a feature comparison to opty:
+
++========+============+================+===========================+======+========================+==========+=============================+
+|        |            |                |                           |      |                        | Implicit |                             |
+| Name   | Language   | License        | Derivatives               | DAEs |  Discretization        | dynamics | Solvers                     |
++========+============+================+===========================+======+========================+==========+=============================+
[…]
