### 2017-10-20 648c7d6
Made table formatting work with markdown rendering.
