### 2017-10-20 csu-hmc/opty#61 @moorepants
I think this is now addressed by #74.
