### 2017-10-24 7c7e9a6
References added to paper.bib
Please check that they appear correctly when paper.md is rendered.  Github does not show me the list of references (how do I get that?).
