### 2017-10-24 csu-hmc/opty#62 @tvdbogert
References were added in paper.bib.  How can I get the rendered version of the list of references?

The comparison table still has gaps.  Which ones should I try to fill?
