### 2017-10-24 csu-hmc/opty#62 @moorepants
I'm not sure how they render the references. I would suggest using a reference manager like jabref, zotero, mendely, etc to export the references to see if they look ok.
