### 2017-10-24 csu-hmc/opty#62 @moorepants
Also, can you help me fill out the table in paper.md. You might know some of the things off hand or where the answer is more quickly.
