### 2017-10-24 4c3e98b
Update paper.md
