### 2017-10-24 5dcaf81
Added von Stryk article to paper.bib
