+++ paper.bib
@@ -18,6 +18,23 @@
 	file = {Snapshot:/home/moorepants/.zotero/zotero/i3casg4l.default/zotero/storage/BXHHBJZW/10.html:text/html}
 }
 
+@Article{vonStryk1992,
+	author="von Stryk, O. and Bulirsch, R.",
+	title="Direct and indirect methods for trajectory optimization",
+	journal="Annals of Operations Research",
+	year="1992",
+	month="Dec",
+	day="01",
+	volume="37",
+	number="1",
+	pages="357--373",
+	abstract="This paper gives a brief list of commonly used direct and indirect efficient methods for the numerical solution of optimal control problems. To improve the low accuracy of the direct methods and to increase the convergence areas of the indirect methods we suggest a hybrid approach. For this a special direct collocation method is presented. In a hybrid approach this direct method can be used in combination with multiple shooting. Numerical examples illustrate the direct method and the hybrid approach.",
+	issn="1572-9338",
+	doi="10.1007/BF02071065",
+	url="https://doi.org/10.1007/BF02071065"
+}
+
+
 @article{vandenBogert2012,
 author = {Antonie J van den Bogert and Maarten Hupperets and Heiko Schlarb and Berthold Krabbe},
 title ={Predictive musculoskeletal simulation using optimal control: effects of added limb mass on energy cost and kinematics of walking and running},
