### 2017-10-24 eb13f9f
Ross article (DIDO reference) added to paper.bib
