+++ paper.bib
@@ -18,6 +18,22 @@
 	file = {Snapshot:/home/moorepants/.zotero/zotero/i3casg4l.default/zotero/storage/BXHHBJZW/10.html:text/html}
 }
 
+@article{Ross2002,
+	title = "A DIRECT METHOD FOR SOLVING NONSMOOTH OPTIMAL CONTROL PROBLEMS",
+	journal = "IFAC Proceedings Volumes",
+	volume = "35",
+	number = "1",
+	pages = "479 - 484",
+	year = "2002",
+	note = "15th IFAC World Congress",
+	issn = "1474-6670",
+	doi = "https://doi.org/10.3182/20020721-6-ES-1901.00329",
+	url = "http://www.sciencedirect.com/science/article/pii/S1474667015387504",
+	author = "I. Michael Ross and Fariba Fahroo",
+	keywords = "Optimal control, discontinuities, discretization, nonlinear programming",
+	abstract = "Abstract We present a class of efficient direct methods for solving nonsmooth dynamic optimization problems where the dynamics are governed by controlled differential inclusions. Our methods are based on pseudospectral approximations of the differential constraints that are assumed to be given in the form of controlled differential inclusions including the usual vector field and differential-algebraic forms. Discontinuities in states, controls, cost functional, dynamic constraints and various other mappings associated with the generalized Bolza problem are allowed by the concept of pseudospectral knots which we introduce in this paper. The computational optimal control problem is reduced to a structured sparse nonlinear programming problem. A simple but illustrative moon-landing problem demonstrates our method."
+}
+
 @Article{vonStryk1992,
 	author="von Stryk, O. and Bulirsch, R.",
 	title="Direct and indirect methods for trajectory optimization",
