### 2017-10-24 89d25cf
Fixed some of the bib file errors.
