+++ paper.bib
@@ -44,49 +44,41 @@ number = {2},
 pages = {123-133},
 year = {2012},
 doi = {10.1177/1754337112440644},
-
-URL = { 
-        https://doi.org/10.1177/1754337112440644
-    
-},
-eprint = { 
-        https://doi.org/10.1177/1754337112440644
-    
-}
-,
-    abstract = { When designing sports equipment, it is often desirable to predict how certain design parameters will affect human performance. In many instances, this requires a consideration of human musculoskeletal mechanics and adaptive neuromuscular control. Current computational methods do not represent these mechanisms, and design optimization typically requires several iterations of prototyping and human testing. This paper introduces a computational method based on musculoskeletal modeling and optimal control, which has the capability to predict the effect of mechanical equipment properties on human performance. The underlying assumption is that users will adapt their neuromuscular control according to an optimality principle, which balances task performance with a minimization of muscular effort. The method was applied to the prediction of metabolic cost and limb kinematics while running and walking with weights attached to the body. A two-dimensional musculoskeletal model was used, with nine kinematic degrees of freedom and 16 muscles. The optimal control problem was solved for two walking speeds and two running speeds, and at each speed, with 200 g and 400 g masses placed at the thigh, knee, shank and foot. The model predicted an increase in energy expenditure that was proportional to the added mass and the effect was largest for a mass placed on the foot. Specifically, the model predicted an energy cost increase of 0.74\% for each 100 g mass added to the foot during running at 3.60 m/s. The model also predicted that stride length would increase by several millimetres in the same condition, relative to the model without added mass. These predictions were consistent with previously published human studies. Peak force and activation remained the same in most muscles, but increased by 26\% in the hamstrings and by 17\% in the rectus femoris for running at 4.27 m/s with 400 g added mass at the foot, suggesting muscle-specific training effects. This work demonstrated that a musculoskeletal model with optimal control can predict the effect of mechanical devices on human performance, and could become a useful tool for design optimization in sports engineering. The theoretical background of predictive simulation also helps explain why human athletes have specific responses when exercising in an altered mechanical environment. }
+URL = { https://doi.org/10.1177/1754337112440644},
+eprint = { https://doi.org/10.1177/1754337112440644},
[…]
