### 2017-10-24 csu-hmc/opty#62 @moorepants
I added a paper.tex file that can be compiled to check the bib entry renderings. I also opened paper.bib in jabref and let it format things and fixed a few errors.
