### 2017-10-24 csu-hmc/opty#62 @tvdbogert
Thanks.  Github's rendering of paper.md ends with the heading "References" and I was expecting to see the list of references there.  I don't even know what a .md file is supposed to do.
