### 2017-10-24 csu-hmc/opty#62 @moorepants
I think that the journal processing the md file and the bib file to create something like: http://joss.theoj.org/papers/050362b7e7691d2a5d0ebed8251bc01e

BTW, that paper is pretty thorough. I though you were just supposed to create an abstract of sorts.
