### 2017-10-24 csu-hmc/opty#62 @tvdbogert
That paper is a couple of paragraphs longer maybe.  I suggest we just address the reviewer comments and not get too ambitious.  As I mentioned earlier, when you try to be thorough, the expectations will rise and you get more criticism.
