### 2017-10-24 csu-hmc/opty#62 @moorepants
Agreed. I think the main remaining things are getting the one example working again (or removing it) and to add some of the mathematical description to the documentation.
