### 2017-10-24 csu-hmc/opty#59 @moorepants
@tvdbogert Would you say this is now addressed?
