+++ paper.tex
@@ -2,7 +2,21 @@
 
 \begin{document}
 
-\cite{Wachter2006,vonStryk1992,vandenBogert2012,Ackermann2010,LeeUmberger2016,LinPandy2017,MooreTGCS2015,Zarei2014,PattersonRao2014,Betts2010,Meurer2017,Behnel2011,Cyipopt2017,Wang2017}
+\cite{Wachter2006,
+      Ross2002,
+      vonStryk1992,
+      vandenBogert2012,
+      Ackermann2010,
+      LeeUmberger2016,
+      LinPandy2017,
+      MooreTGCS2015,
+      Zarei2014,
+      PattersonRao2014,
+      Betts2010,
+      Meurer2017,
+      Behnel2011,
+      Cyipopt2017,
+      Wang2017}
 
 \bibliographystyle{plain}
 \bibliography{paper.bib}
