### 2017-10-24 csu-hmc/opty#59 @tvdbogert
Yes, but there are still a few gaps in the table (in the rows you added).
