### 2017-10-24 9908ccb
First shot at the theory. Fixes #60.
