+++ docs/index.rst
@@ -12,7 +12,7 @@ Contents
 .. toctree::
    :maxdepth: 2
 
-   index
+   theory
    examples
    api
 
