### 2017-10-24 csu-hmc/opty#59 @moorepants
This specific review comment didn't ask for the table, so I'm going to close this one.
