### 2017-10-24 2db9de1
More updates to theory.
