### 2017-10-24 a6cd1f6
Some updates to the table.
