+++ paper.md
@@ -76,34 +76,35 @@ There are existing software packages that have similarities to opty. Below, is
 a feature comparison to opty:
 
 ```
-+========+============+================+===========================+======+========================+==========+=============================+
-|        |            |                |                           |      |                        | Implicit |                             |
-| Name   | Language   | License        | Derivatives               | DAEs |  Discretization        | dynamics | Solvers                     | URL |
-+========+============+================+===========================+======+========================+==========+=============================+
-| DIRCOL | Fortran    | Non-commercial | Finite differences        | Yes  | Piecewise linear/cubic | Yes      | NPSOL, SNOPT                            |
-+--------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| GPOPS  | Matlab     | Commercial     | Automatic differentiation | No   | Pseudospectral         | No       | SNOPT, IPOPT                       |
-+--------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| SOCS   | Fortran    | Commercial     | Finite differences        | Yes  | Euler, RK, & others    | Yes      | built-in                            |
-+--------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| PROPT  | Matlab     | Commercial     | Analytic                  | Yes  | Pseudospectral         | Yes      | SNOPT, KNITRO                            |
-+--------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| DIDO   | Matlab     | Commercial     | Analytic                  | No   | Pseudospectral         | Yes      | built-in                            |
-+--------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| Casadi | C++/Python | LGPL           | Automatic differentiation | Yes  | ?                      | Yes      | IPOPT, SNOPT, WORHP, KNITRO |
-+--------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| PSOPT  | C++        | GPL            | Automatic differentiation |      | Pseudospectral         |          | IPOPT, SNOPT                |
-|        |            |                | Sparse finite differences |      |                        |          |                             |
-+--------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| GESOP  |            |                |                           |      |                        |          |                             |
-+--------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| DYNOPT |            |                |                           |      |                        |          |                             |
-+--------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| opty   | Python     | BSD 2-Clause   | Analytic                  | ?    | Euler, Midpoint        |          | IPOPT                       |
-+--------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
++========+=====================+============+================+===========================+======+========================+==========+=============================+
+|        |                     |            |                |                           |      |                        | Implicit |                             |
+| Name   | Citation            | Language   | License        | Derivatives               | DAEs |  Discretization        | dynamics | Solvers                     |
++========+=====================+============+===========================+======+========================+==========+==============================================+
+| Casadi |                     | C++/Python | LGPL           | Automatic differentiation | Yes  | ?                      | Yes      | IPOPT, SNOPT, WORHP, KNITRO |
++--------+---------------------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
[…]
