### 2017-10-25 fbef235
Improvements to the paper.
- Including Casadi even thought it doesn't have the direct collocation
  built in, but it can solve dc problems, as their examples show.
- Inlcuded a url column in the table.
- Turned opty's strengths into a paragraph.
