+++ paper.bib
@@ -246,4 +246,28 @@ abstract = {Solving an optimal control or estimation problem is not easy. Pieces
   url       = {http://ruina.tam.cornell.edu/dynwalk17abstracts/258-Anne%20Koelewijn%20-%20HuaweiAbstract.pdf},
 }
 
+@PHDTHESIS{Andersson2013,
+  author = {Joel Andersson},
+  title = {{A} {G}eneral-{P}urpose {S}oftware {F}ramework for {D}ynamic {O}ptimization},
+  school = {Arenberg Doctoral School, KU Leuven},
+  year = {2013},
+  type = {{P}h{D} thesis},
+  address = {Department of Electrical Engineering (ESAT/SCD) and Optimization in Engineering Center,
+             Kasteelpark Arenberg 10, 3001-Heverlee, Belgium},
+  month = {October}
+}
+
+@inproceedings{Becerra2010,
+	title = {Solving complex optimal control problems at no cost with {PSOPT}},
+	doi = {10.1109/CACSD.2010.5612676},
+	abstract = {This paper introduces PSOPT, an open source optimal control solver written in C++. PSOPT uses pseudospectral and local discretizations, sparse nonlinear programming, automatic differentiation, and it incorporates automatic scaling and mesh refinement facilities. The software is able to solve complex optimal control problems including multiple phases, delayed differential equations, nonlinear path constraints, interior point constraints, integral constraints, and free initial and/or final times. The software does not require any non-free platform to run, not even the operating system, as it is able to run under Linux. Additionally, the software generates plots as well as LATEX code so that its results can easily be included in publications. An illustrative example is provided.},
+	booktitle = {2010 {IEEE} {International} {Symposium} on {Computer}-{Aided} {Control} {System} {Design}},
+	author = {Becerra, V. M.},
+	month = sep,
+	year = {2010},
+	keywords = {automatic differentiation, C++, C++ language, Chebyshev approximation, complex optimal control problems solving, delayed differential equations, differential equations, Differential equations, integral constraints, interior point constraints, Interpolation, LATEX code, Linux, mesh refinement facilities, multiple phases, nonlinear path constraints, nonlinear programming, open source optimal control solver, optimal control, Optimal control, Polynomials, PSOPT, public domain software, Software, sparse nonlinear programming},
+	pages = {1391--1396},
+	file = {IEEE Xplore Abstract Record:/home/moorepants/Zotero/storage/7PNMFTS8/5612676.html:text/html;IEEE Xplore Full Text PDF:/home/moorepants/Zotero/storage/FXZAZDHP/Becerra - 2010 - Solving complex optimal control problems at no cos.pdf:application/pdf}
+}
+
 @Comment{jabref-meta: databaseType:bibtex;}
