+++ paper.md
@@ -72,62 +72,55 @@ however, the low order discretizations provide more robust convergence to the
 globally optimal trajectory in a nonlinear system when a good initial guess is
 not available [@Zarei2016].
 
-There are existing software packages that have similarities to opty. Below, is
-a feature comparison to opty:
+There are existing software packages that can solve optimal control problems
+and have have similarities to opty. Below, is a feature comparison:
 
 ```
-+========+=====================+============+================+===========================+======+========================+==========+=============================+
-|        |                     |            |                |                           |      |                        | Implicit |                             |
-| Name   | Citation            | Language   | License        | Derivatives               | DAEs |  Discretization        | dynamics | Solvers                     |
-+========+=====================+============+===========================+======+========================+==========+==============================================+
-| Casadi |                     | C++/Python | LGPL           | Automatic differentiation | Yes  | ?                      | Yes      | IPOPT, SNOPT, WORHP, KNITRO |
-+--------+---------------------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| DIDO   | [@Ross2002]         | Matlab     | Commercial     | Analytic                  | No   | Pseudospectral         | Yes      | built-in                    |
-+--------+---------------------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| DIRCOL | [@vonStryk1993]     | Fortran    | Non-commercial | Finite differences        | Yes  | Piecewise linear/cubic | Yes      | NPSOL, SNOPT                |
-+--------+---------------------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| DYNOPT |                     |            |                |                           |      |                        |          |                             |
-+--------+---------------------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| GESOP  |                     |            |                |                           |      |                        |          |                             |
-+--------+---------------------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| GPOPS  | [@PattersonRao2014] | Matlab     | Commercial     | Automatic differentiation | No   | Pseudospectral         | No       | SNOPT, IPOPT                |
-+--------+---------------------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| opty   |                     | Python     | BSD 2-Clause   | Analytic                  | Yes  | Euler, Midpoint        | Yes      | IPOPT                       |
-+--------+---------------------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| PROPT  |                     | Matlab     | Commercial     | Analytic                  | Yes  | Pseudospectral         | Yes      | SNOPT, KNITRO               |
-+--------+---------------------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| PSOPT  |                     | C++        | GPL            | Automatic differentiation |      | Pseudospectral         |          | IPOPT, SNOPT                |
-|        |                     |            |                | Sparse finite differences |      |                        |          |                             |
-+--------+---------------------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
-| SOCS   | [@Betts2010]        | Fortran    | Commercial     | Finite differences        | Yes  | Euler, RK, & others    | Yes      | built-in                    |
-+--------+---------------------+------------+----------------+---------------------------+------+------------------------+----------+-----------------------------+
++========+=====================+============+================+=============================+======+========================+==========+=================+=================================================================================================+
+|        |                     |            |                |                             |      |                        | Implicit |                 |                                                                                                 |
[…]
