+++ paper.md
@@ -96,7 +96,7 @@ and have have similarities to opty. Below, is a feature comparison:
 +--------+---------------------+------------+----------------+-----------------------------+------+------------------------+----------+-----------------+-------------------------------------------------------------------------------------------------+
 | opty   |                     | Python     | BSD 2-Clause   | Analytic                    | Yes  | Euler, Midpoint        | Yes      | IPOPT           | `opty Documentation <http://opty.readthedocs.io>`_                                              |
 +--------+---------------------+------------+----------------+-----------------------------+------+------------------------+----------+-----------------+-------------------------------------------------------------------------------------------------+
-| PROPT  |                     | Matlab     | Commercial     | Analytic                    | Yes  | Pseudospectral         | Yes      | SNOPT, KNITRO   | `TOMDYN Website <http://tomdyn.com/index.html>`_                                                |
+| PROPT  | [@Rutquist2010]     | Matlab     | Commercial     | Analytic                    | Yes  | Pseudospectral         | Yes      | SNOPT, KNITRO   | `TOMDYN Website <http://tomdyn.com/index.html>`_                                                |
 +--------+---------------------+------------+----------------+-----------------------------+------+------------------------+----------+-----------------+-------------------------------------------------------------------------------------------------+
 | PSOPT  | [@Becerra2010]      | C++        | GPL            | - Automatic differentiation |      | Pseudospectral         |          | IPOPT, SNOPT    |                                                                                                 |
 |        |                     |            |                | - Sparse finite differences |      |                        |          |                 |                                                                                                 |
