+++ paper.bib
@@ -1,68 +1,68 @@
 % Encoding: UTF-8
 
 @article{Wachter2006,
-	title = {On the implementation of an interior-point filter line-search algorithm for large-scale nonlinear programming},
-	volume = {106},
-	issn = {0025-5610, 1436-4646},
-	url = {https://link.springer.com/article/10.1007/s10107-004-0559-y},
-	doi = {10.1007/s10107-004-0559-y},
-	abstract = {.We present a primal-dual interior-point algorithm with a filter line-search method for nonlinear programming. Local and global convergence properties of this method were analyzed in previous work. Here we provide a comprehensive description of the algorithm, including the feasibility restoration phase for the filter method, second-order corrections, and inertia correction of the KKT matrix. Heuristics are also considered that allow faster performance. This method has been implemented in the IPOPT code, which we demonstrate in a detailed numerical study based on 954 problems from the CUTEr test set. An evaluation is made of several line-search options, and a comparison is provided with two state-of-the-art interior-point codes for nonlinear programming.},
-	language = {en},
-	number = {1},
-	urldate = {2017-06-04},
-	journal = {Mathematical Programming},
-	author = {Wächter, Andreas and Biegler, Lorenz T.},
-	month = mar,
-	year = {2006},
-	pages = {25--57},
-	file = {Snapshot:/home/moorepants/.zotero/zotero/i3casg4l.default/zotero/storage/BXHHBJZW/10.html:text/html}
+  title = {On the implementation of an interior-point filter line-search algorithm for large-scale nonlinear programming},
+  volume = {106},
+  issn = {0025-5610, 1436-4646},
+  url = {https://link.springer.com/article/10.1007/s10107-004-0559-y},
+  doi = {10.1007/s10107-004-0559-y},
+  abstract = {.We present a primal-dual interior-point algorithm with a filter line-search method for nonlinear programming. Local and global convergence properties of this method were analyzed in previous work. Here we provide a comprehensive description of the algorithm, including the feasibility restoration phase for the filter method, second-order corrections, and inertia correction of the KKT matrix. Heuristics are also considered that allow faster performance. This method has been implemented in the IPOPT code, which we demonstrate in a detailed numerical study based on 954 problems from the CUTEr test set. An evaluation is made of several line-search options, and a comparison is provided with two state-of-the-art interior-point codes for nonlinear programming.},
+  language = {en},
+  number = {1},
+  urldate = {2017-06-04},
+  journal = {Mathematical Programming},
+  author = {Wächter, Andreas and Biegler, Lorenz T.},
+  month = mar,
+  year = {2006},
+  pages = {25--57},
+  file = {Snapshot:/home/moorepants/.zotero/zotero/i3casg4l.default/zotero/storage/BXHHBJZW/10.html:text/html}
 }
 
 @article{Ross2002,
-	title = "A DIRECT METHOD FOR SOLVING NONSMOOTH OPTIMAL CONTROL PROBLEMS",
-	journal = "IFAC Proceedings Volumes",
-	volume = "35",
-	number = "1",
-	pages = "479 - 484",
-	year = "2002",
-	note = "15th IFAC World Congress",
-	issn = "1474-6670",
-	doi = "https://doi.org/10.3182/20020721-6-ES-1901.00329",
-	url = "http://www.sciencedirect.com/science/article/pii/S1474667015387504",
-	author = "I. Michael Ross and Fariba Fahroo",
-	keywords = "Optimal control, discontinuities, discretization, nonlinear programming",
[…]
