### 2017-10-25 388eb48
Completed the table for all but two.
