+++ paper.md
@@ -80,25 +80,25 @@ and have have similarities to opty. Below, is a feature comparison:
 |        |                     |            |                |                             |      |                        | Implicit |                 |                                                                                                 |
 | Name   | Citation            | Language   | License        | Derivatives                 | DAEs |  Discretization        | dynamics | Solvers         | Project Website                                                                                 |
 +========+=====================+============+================+=============================+======+========================+==========+=================+=================================================================================================+
-| Casadi | [@Andersson2013]    | - C++      | LGPL           | Automatic differentiation   | Yes  | None                   | Yes      | - IPOPT, WORHP  |                                                                                                 |
+| Casadi | [@Andersson2013]    | - C++      | LGPL           | Automatic differentiation   | Yes  | None                   | Yes      | - IPOPT, WORHP  | `Casadi Website <https://github.com/casadi/casadi/wiki>`_                                       |
 |        |                     | - Python   |                |                             |      |                        |          | - SNOPT, KNITRO |                                                                                                 |
 |        |                     | - Octave   |                |                             |      |                        |          |                 |                                                                                                 |
 +--------+---------------------+------------+----------------+-----------------------------+------+------------------------+----------+-----------------+-------------------------------------------------------------------------------------------------+
-| DIDO   | [@Ross2002]         | Matlab     | Commercial     | Analytic                    | No   | Pseudospectral         | Yes      | built-in        | `DIDO Website <http://www.elissarglobal.com/industry/products/software-3/)>`_                   |
+| DIDO   | [@Ross2002]         | Matlab     | Commercial     | Analytic                    | No   | Pseudospectral         | Yes      | built-in        | `DIDO Website <http://www.elissarglobal.com/industry/products/software-3/>`_                    |
 +--------+---------------------+------------+----------------+-----------------------------+------+------------------------+----------+-----------------+-------------------------------------------------------------------------------------------------+
-| DIRCOL | [@vonStryk1993]     | Fortran    | Non-commercial | Finite differences          | Yes  | Piecewise linear/cubic | Yes      | NPSOL, SNOPT    | `DIRCOL Website <http://www.sim.informatik.tu-darmstadt.de/en/res/sw/dircol/)>`_                |
+| DIRCOL | [@vonStryk1993]     | Fortran    | Non-commercial | Finite differences          | Yes  | Piecewise linear/cubic | Yes      | NPSOL, SNOPT    | `DIRCOL Website <http://www.sim.informatik.tu-darmstadt.de/en/res/sw/dircol/>`_                 |
 +--------+---------------------+------------+----------------+-----------------------------+------+------------------------+----------+-----------------+-------------------------------------------------------------------------------------------------+
 | DYNOPT |                     |            |                |                             |      |                        |          |                 |                                                                                                 |
 +--------+---------------------+------------+----------------+-----------------------------+------+------------------------+----------+-----------------+-------------------------------------------------------------------------------------------------+
 | GESOP  |                     |            |                |                             |      |                        |          |                 |                                                                                                 |
 +--------+---------------------+------------+----------------+-----------------------------+------+------------------------+----------+-----------------+-------------------------------------------------------------------------------------------------+
-| GPOPS  | [@PattersonRao2014] | Matlab     | Commercial     | Automatic differentiation   | No   | Pseudospectral         | No       | SNOPT, IPOPT    | `GPOPS Website <http://www.gpops2.com/)>`_                                                      |
[…]
