### 2017-10-25 csu-hmc/opty#62 @moorepants
@tvdbogert We are only missing information for DYNOPT and GESOP. Do you have info or should I research them?

I also just found this: https://otis.grc.nasa.gov/
