### 2017-10-26 674e5bf
More work on the theory, fixed errors, clarified.
