+++ docs/theory.rst
@@ -2,7 +2,7 @@
 Theory
 ======
 
-Given at a set of continous differential equations in implicit form:
+Given at a set of continuous differential equations in time in implicit form:
 
 .. math::
 
@@ -15,83 +15,113 @@ where:
 - :math:`\mathbf{p}` is the vector of constant parameters
 - :math:`t` is time
 
-one can then break up :math:`\mathbf{r}` and :math:`\mathbf{p}` into known and
-unknown quantities:
+one can then break up :math:`\mathbf{r}` and :math:`\mathbf{p}` into known $k$
+and unknown $u$ quantities:
 
 .. math::
 
-   \mathbf{r} = \left[ \mathbf{r}_k \quad \mathbf{r}_u \right]
+   \mathbf{r}(t) = \left[ \mathbf{r}_k(t) \quad \mathbf{r}_u(t) \right]^T
 
-   \mathbf{r} = \left[ \mathbf{p}_k \quad \mathbf{p}_u \right]
+   \mathbf{p} = \left[ \mathbf{p}_k \quad \mathbf{p}_u \right]^T
 
 Then there are optimal state trajectories :math:`\mathbf{x}`, optimal input
 trajectories :math:`\mathbf{r}_u` and optimal parameter values
 :math:`\mathbf{p}_u` if some cost function, :math:`J(\mathbf{x}, \mathbf{r}_u,
-\mathbf{p})` is specified. In addition one may specify some constraints on the
-trajectories or parameters in the form of:
+\mathbf{p})` is specified and any optional constraints on these variables in
+the form of:
 
 .. math::
 
-    \mathbf{c}^L \leq \mathbf{c} \leq \mathbf{c}^U
+   \mathbf{x}^L \leq \mathbf{x} \leq \mathbf{x}^U \\
+   \mathbf{u}^L \leq \mathbf{u} \leq \mathbf{u}^U \\
+   \mathbf{p}^L \leq \mathbf{p} \leq \mathbf{p}^U
 
-Constraints can be:
+Constraints are commonly:
 
 - path constraints
 - bounds on the trajectories or parameter values
 - specific trajectory or parameter values at instances of time
 
-This is typically called an optimal control problem where you are seeking the
-optimal open loop input trajectories and/or the optimal system parameters such
-that the dynamical system does a task in an optimal way. This problem can be
-rewritten as a non-linear programming problem using direct collocation methods
-and solved using optimizers for NLP problems.
+This type of problem called an `optimal control`_ problem with the objective of
+finding the open loop input trajectories and/or the optimal system parameters
+such that the dynamical system evolves in time in a way that meets the
+objective and constraints. This problem can be rewritten as a `nonlinear
+programming`_ (NLP) problem using `direct collocation`_ methods and solved
+using a variety of NLP optimization algorithms.
 
-In addition, to the above constraints the differential equations can be
-discretized using a variety of integration methods and appended as additional
-optimality constraints. This ensures that the dynamics are abided by at each
-discretized time instance and relieves the need to integrate the equations as
-one does in shooting methods. Using the backward Euler method of integration,
-the approximation of the derivative of the states is:
+.. _optimal control: https://en.wikipedia.org/wiki/Optimal_control
+.. _nonlinear programming: https://en.wikipedia.org/wiki/Nonlinear_programming
[…]
