### 2017-10-26 552407d
Updated theory wrt to Ton's comments.
