+++ docs/theory.rst
@@ -6,34 +6,67 @@ Given at a set of continuous differential equations in time in implicit form:
 
 .. math::
 
-   \mathbf{f}(\dot{\mathbf{x}}(t), \mathbf{x}(t), \mathbf{r}(t), \mathbf{p}, t) = \mathbf{0}
+   \mathbf{f}(\dot{\mathbf{y}}(t), \mathbf{y}(t), \mathbf{r}(t), \mathbf{p}, t) = \mathbf{0}
 
 where:
 
-- :math:`\mathbf{x}` is the state vector
-- :math:`\mathbf{r}` is the vector of specified inputs
-- :math:`\mathbf{p}` is the vector of constant parameters
-- :math:`t` is time
+- :math:`t` is continuous time
+- :math:`\mathbf{y}(t) \in \mathbb{R}^n` the state vector at time
+  :math:`t`
+- :math:`\mathbf{r}(t) \in \mathbb{R}^p` is the vector of specified
+  inputs at time :math:`t`
+- :math:`\mathbf{p}(t) \in \mathbb{R}^q` is the vector of constant parameters
 
-one can then break up :math:`\mathbf{r}` and :math:`\mathbf{p}` into known $k$
-and unknown $u$ quantities:
+From here on out, :math:`(t)` will be dropped for convenience.
+
+In opty, you would define these equations as a SymPy column matrix that
+contains SymPy expressions. For example, a simple compound pendulum is
+described by the first order ordinary differential equations:
 
 .. math::
 
-   \mathbf{r}(t) = \left[ \mathbf{r}_k(t) \quad \mathbf{r}_u(t) \right]^T
+   \begin{bmatrix}
+   \dot{\theta} - \omega\\
+   I \dot{\omega} + mgl\sin\theta - T
+   \end{bmatrix}
+   = \mathbf{0}
+
+In SymPy, this would look like:
+
+.. code:: pycon
+
+   >>> I, m, g, l, t = symbols('I, m, g, l, t')
+   >>> theta, omega, T = symbols('theta, omega, T', cls=Function)
+   >>> f = Matrix([theta(t).diff() - omega(t),
+   ...             I*omega(t).diff() + m*g*l*sin(theta(t)) - T(t)])
+
+One can then break up :math:`\mathbf{r}` and :math:`\mathbf{p}` into known,
+:math:`k`, and unknown, :math:`u`, quantities.
+
+.. math::
+
+   \mathbf{r} = \left[ \mathbf{r}_k \quad \mathbf{r}_u \right]^T
 
    \mathbf{p} = \left[ \mathbf{p}_k \quad \mathbf{p}_u \right]^T
 
-Then there are optimal state trajectories :math:`\mathbf{x}`, optimal input
+Then there are optimal state trajectories :math:`\mathbf{y}`, optimal input
 trajectories :math:`\mathbf{r}_u` and optimal parameter values
-:math:`\mathbf{p}_u` if some cost function, :math:`J(\mathbf{x}, \mathbf{r}_u,
-\mathbf{p})` is specified and any optional constraints on these variables in
-the form of:
+:math:`\mathbf{p}_u` if some cost function, :math:`J(\mathbf{y}, \mathbf{r}_u,
+\mathbf{p}_u)` is specified, for example, the integral of a cost rate :math:`L`
+with respect to time:
 
 .. math::
 
-   \mathbf{x}^L \leq \mathbf{x} \leq \mathbf{x}^U \\
-   \mathbf{u}^L \leq \mathbf{u} \leq \mathbf{u}^U \\
+   J(\mathbf{y}, \mathbf{r}_u, \mathbf{p}_u) =
[…]
