### 2017-10-26 e0578cf
Added NCSRR funding to README.
