### 2017-10-26 #300 @moorepants
A couple of questions:

- What is the recommended way to put tables in the paper.md file?
- How do we check that the bibliography renders correctly?
