### 2017-10-26 #300 @arfon
@moorepants - the command we run is here: https://github.com/openjournals/whedon/blob/master/lib/whedon/processor.rb#L173

> How do we check that the bibliography renders correctly?

Looks like there's a small issue with your bibliography:

```
pandoc-citeproc: reference Watcher2006 not found
```

> What is the recommended way to put tables in the paper.md file?

I'm not sure sorry but the current table looks a bit wonky: [10.21105.joss.00300.pdf](https://github.com/openjournals/joss-reviews/files/1420197/10.21105.joss.00300.pdf)

Do any of the @openjournals/joss-editors know how to fix this?
