### 2017-10-27 #300 @moorepants
I wasn't sure if there was a spec for making tables in markdown so I made an restructuredtext table and just wrapped in in markdown literal ticks for the time being. I knew it would render as is.

Is there a simple way to run that command on my computer? I'm not familiar with Ruby.
