### 2017-10-27 #300 @Kevin-Mattheus-Moerman
@arfon obviously we could use [Markdown](https://github.com/adam-p/markdown-here/wiki/Markdown-Cheatsheet#tables ), [HTML](https://www.w3schools.com/html/html_tables.asp), or [LATEX](https://en.wikibooks.org/wiki/LaTeX/Tables) tables. But I'm not sure how the process of creating the PDF handles these. Does normal Markdown table stuff work?

| Tables        | Are           | Cool  |
| ------------- |:-------------:| -----:|
| col 3 is      | right-aligned | Nice |
| col 2 is      | centered      |   Fits well |
| zebra stripes | are neat      |    3.14159265359 |
