### 2017-10-27 #300 @moorepants
So Github flavored markdown has the table support that you sow above: https://help.github.com/articles/organizing-information-with-tables/, but this isn't in all markdown flavors. It isn't clear to me in the ruby code which flavor of markdown is being used. Pandoc can process a number of them apparently.
