### 2017-10-28 #300 @arfon
> Is there a simple way to run that command on my computer? I'm not familiar with Ruby.

```
      `cd #{paper_directory} && pandoc \
      -V repository="#{repository_address}" \
      -V archive_doi="#{archive_doi}" \
      -V paper_url="#{paper_url}" \
      -V formatted_doi="#{formatted_doi}" \
      -V review_issue_url="#{review_issue_url}" \
      -V graphics="true" \
      -V issue="#{paper_issue}" \
      -V volume="#{paper_volume}" \
      -V page="#{review_issue_id}" \
      -V joss_logo_path="#{Dir.pwd}/resources/joss-logo.png" \
      -V year="#{paper_year}" \
      -V formatted_doi="#{formatted_doi}" \
      -V citation_author="#{citation_author}" \
      -V paper_title="#{paper_title}" \
      -S -o #{filename_doi}.pdf -V geometry:margin=1in \
      --latex-engine=xelatex \
      --filter pandoc-citeproc #{File.basename(paper_path)} \
      --template #{latex_template_path}
```

The command above is just passing a bunch of variables in but it's a normal(ish) pandoc command so if you have pandoc installed locally you should be able to do something like:

```
pandoc  -S -o paper.pdf --latex-engine=xelatex --filter pandoc-citeproc paper.md --template latex.template
```
...where the `latex.template` is [this file](https://github.com/openjournals/whedon/blob/master/resources/latex.template). I'm also happy to try compiling the document a few times with you if you want to debug this here. In the future we plan to give @whedon the ability to compile the PDF on demand for you.
