### 2017-11-07 8e14041
Made openmp parallelization optional with check for openmp.
