### 2017-11-07 6af7683
Added an example showing speed comparisons for parallel execution.
