### 2017-11-07 60f6788
Test openmp compilation on travis.
