+++ .travis.yml
@@ -18,7 +18,7 @@ install:
   - conda info -a
   - conda create -q -n test-environment python=$TRAVIS_PYTHON_VERSION
   - source activate test-environment
-  - conda install sympy cython cyipopt nose coverage sphinx matplotlib
+  - conda install sympy cython cyipopt nose coverage sphinx matplotlib openmp
 script:
   - nosetests -v --with-coverage --cover-package=opty
   - python setup.py install
