### 2017-11-07 4ce12a2
Added user warning if parallel request is ignored.
