### 2017-11-21 csu-hmc/opty#62 @moorepants
Just came across this. It says is had symbolics and trajectory optimization: http://ayonga.github.io/frost-dev/pages/introduction.html
