### 2017-11-22 csu-hmc/opty#62 @moorepants
BTW, I'm unlikely to finish this until xmas break. I've been working 60+ hr weeks to keep up with classes at the moment...
