### 2017-12-04 2dd655f
Fixed some math.
