### 2018-01-08 #300 @Kevin-Mattheus-Moerman
@stavness Are you able to start ticking some of the boxes at the top of this review issue?
