### 2018-01-08 #300 @stavness
I'm waiting for the revised version, and then I'll look at that time.

On Mon, Jan 8, 2018 at 10:33 AM, Kevin Mattheus Moerman <
notifications@github.com> wrote:

> @stavness <https://github.com/stavness> Are you able to start ticking
> some of the boxes at the top of this review issue?
>
