### 2018-01-08 #300 @moorepants
I apologize, my revision stalled due to other priorities. This is near the top of my task list, so I except to finalize this month.
