### 2018-01-09 5979876
More improvements to the theory doc.
