+++ docs/theory.rst
@@ -2,22 +2,23 @@
 Theory
 ======
 
-Given at a set of continuous differential equations in time in implicit form:
+Given at a set of first order continuous differential equations in time in
+implicit form:
 
 .. math::
 
-   \mathbf{f}(\dot{\mathbf{y}}(t), \mathbf{y}(t), \mathbf{r}(t), \mathbf{p}, t) = \mathbf{0}
+   \mathbf{f}( \dot{\mathbf{y}}(t), \mathbf{y}(t), \mathbf{r}(t), \mathbf{p}, t ) = \mathbf{0}
 
 where:
 
-- :math:`t` is continuous time
+- :math:`t` is time
 - :math:`\mathbf{y}(t) \in \mathbb{R}^n` the state vector at time
   :math:`t`
 - :math:`\mathbf{r}(t) \in \mathbb{R}^p` is the vector of specified
-  inputs at time :math:`t`
+  (exongenous) inputs at time :math:`t`
 - :math:`\mathbf{p}(t) \in \mathbb{R}^q` is the vector of constant parameters
 
-From here on out, :math:`(t)` will be dropped for convenience.
+From here on out, the notation :math:`(t)` will be dropped for convenience.
 
 In opty, you would define these equations as a SymPy column matrix that
 contains SymPy expressions. For example, a simple compound pendulum is
@@ -25,8 +26,13 @@ described by the first order ordinary differential equations:
 
 .. math::
 
+   \mathbf{f}\left(\begin{bmatrix}\dot{\theta} \\ \dot{\omega}\end{bmatrix},
+   \begin{bmatrix}\theta \\ \omega\end{bmatrix},
+   \begin{bmatrix}T\end{bmatrix},
+   \begin{bmatrix}I \\ m \\ g \\ l \end{bmatrix}\right)
+   =
    \begin{bmatrix}
-   \dot{\theta} - \omega\\
+   \dot{\theta} - \omega \\
    I \dot{\omega} + mgl\sin\theta - T
    \end{bmatrix}
    = \mathbf{0}
@@ -49,19 +55,19 @@ One can then break up :math:`\mathbf{r}` and :math:`\mathbf{p}` into known,
 
    \mathbf{p} = \left[ \mathbf{p}_k \quad \mathbf{p}_u \right]^T
 
-Then there are optimal state trajectories :math:`\mathbf{y}`, optimal input
-trajectories :math:`\mathbf{r}_u` and optimal parameter values
+Then there are optimal state trajectories :math:`\mathbf{y}`, optimal unknown
+input trajectories :math:`\mathbf{r}_u` and optimal unknown parameter values
 :math:`\mathbf{p}_u` if some cost function, :math:`J(\mathbf{y}, \mathbf{r}_u,
-\mathbf{p}_u)` is specified, for example, the integral of a cost rate :math:`L`
+\mathbf{p}_u)` is specified. For example, the integral of a cost rate :math:`L`
 with respect to time:
 
 .. math::
 
    J(\mathbf{y}, \mathbf{r}_u, \mathbf{p}_u) =
-   \int L(\mathbf{x}(t), \mathbf{r}(t), \mathbf{p}) dt
+   \int L(\mathbf{y}(t), \mathbf{r}(t), \mathbf{p}) dt
 
-Additionally, any optional constraints on these variables can be specified the
-form of:
+Additionally, upper :math:`U` and lower :math:`L` boundary constraints on these
+variables can be specified as follows:
 
 .. math::
 
@@ -69,39 +75,41 @@ form of:
    \mathbf{r}^L \leq \mathbf{r} \leq \mathbf{r}^U \\
    \mathbf{p}^L \leq \mathbf{p} \leq \mathbf{p}^U
 
-Constraints are commonly:
[…]
