### 2018-01-09 csu-hmc/opty#60 @moorepants
I've fixed this one. The docs are now rendered here: http://opty.readthedocs.io/en/latest/theory.html
