### 2018-01-09 e06cb9d
Added sympy output.
