### 2018-01-09 c9213f9
Corrected a typo in theory.rst
In problem description, p(t) changed to p (because it is defined as a vector with constant parameters.
