+++ paper.md
@@ -75,34 +75,37 @@ not available [@Zarei2016].
 There are existing software packages that can solve optimal control problems
 and have have similarities to opty. Below, is a feature comparison:
 
++--------+---------------------+------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
+| Name   | Citation            | Language   | License        | Derivatives                 | DAEs |  Discretization        | Implicit Dynamics | Solvers         | Project Website                                                                                 |
++========+=====================+============+================+=============================+======+========================+===================+=================+=================================================================================================+
+| Casadi | [@Andersson2013]    | C++,       | LGPL           | Automatic differentiation   | Yes  | None                   | Yes               | IPOPT, WORHP,   | `Casadi Website <https://github.com/casadi/casadi/wiki>`_                                       |
+| [1]_   |                     | Python,    |                |                             |      |                        |                   | SNOPT, KNITRO   |                                                                                                 |
+|        |                     | Octave,    |                |                             |      |                        |                   |                 |                                                                                                 |
++--------+---------------------+------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
+| DIDO   | [@Ross2002]         | Matlab     | Commercial     | Analytic                    | No   | Pseudospectral         | Yes               | built-in        | `DIDO Website <http://www.elissarglobal.com/industry/products/software-3/>`_                    |
++--------+---------------------+------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
+| DIRCOL | [@vonStryk1993]     | Fortran    | Non-commercial | Finite differences          | Yes  | Piecewise linear/cubic | Yes               | NPSOL, SNOPT    | `DIRCOL Website <http://www.sim.informatik.tu-darmstadt.de/en/res/sw/dircol/>`_                 |
++--------+---------------------+------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
+| DYNOPT | [2]_                |            |                |                             |      |                        |                   |                 |                                                                                                 |
++--------+---------------------+------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
+| GESOP  | [2]_                |            |                |                             |      |                        |                   |                 |                                                                                                 |
++--------+---------------------+------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
+| GPOPS  | [@PattersonRao2014] | Matlab     | Commercial     | Automatic differentiation   | No   | Pseudospectral         | No                | SNOPT, IPOPT    | `GPOPS Website <http://www.gpops2.com/>`_                                                       |
++--------+---------------------+------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
+| opty   | NA                  | Python     | BSD 2-Clause   | Analytic                    | Yes  | Euler, Midpoint        | Yes               | IPOPT           | `opty Documentation <http://opty.readthedocs.io>`_                                              |
++--------+---------------------+------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
+| PROPT  | [@Rutquist2010]     | Matlab     | Commercial     | Analytic                    | Yes  | Pseudospectral         | Yes               | SNOPT, KNITRO   | `TOMDYN Website <http://tomdyn.com/index.html>`_                                                |
[…]
