### 2018-01-10 b69bf74
Fixed literal bounds.
