### 2018-01-10 9d23078
Replaced rst table in markdown with HTML.
