+++ paper.bib
@@ -279,4 +279,15 @@ Key          = {}
   volume      = {260},
 }
 
+@inproceedings{Hereid2017,
+  title = {FROST: Fast Robot Optimization and Simulation Toolkit},
+  author = {Hereid, Ayonga and Ames, Aaron D.},
+  booktitle = {IEEE/RSJ International Conference on Intelligent Robots and Systems (IROS)},
+  year = {2017},
+  address = {Vancouver, BC, Canada},
+  month = sep,
+  organization = {IEEE/RSJ},
+  month_numeric = {9}
+}
+
 @Comment{jabref-meta: databaseType:bibtex;}
