+++ paper-table.rst
@@ -1,0 +1,37 @@
+
++-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
+| Name        | Citation            | Language    | License        | Derivatives                 | DAEs |  Discretization        | Implicit Dynamics | Solvers         | Project Website                                                                                 |
++=============+=====================+=============+================+=============================+======+========================+===================+=================+=================================================================================================+
+| Casadi [1]_ | [@Andersson2013]    | C++,        | LGPL           | Automatic differentiation   | Yes  | None                   | Yes               | IPOPT, WORHP,   | `Casadi Website <https://github.com/casadi/casadi/wiki>`_                                       |
+|             |                     | Python,     |                |                             |      |                        |                   | SNOPT, KNITRO   |                                                                                                 |
+|             |                     | Octave,     |                |                             |      |                        |                   |                 |                                                                                                 |
++-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
+| DIDO        | [@Ross2002]         | Matlab      | Commercial     | Analytic                    | No   | Pseudospectral         | Yes               | built-in        | `DIDO Website <http://www.elissarglobal.com/industry/products/software-3/>`_                    |
++-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
+| DIRCOL      | [@vonStryk1993]     | Fortran     | Non-commercial | Finite differences          | Yes  | Piecewise linear/cubic | Yes               | NPSOL, SNOPT    | `DIRCOL Website <http://www.sim.informatik.tu-darmstadt.de/en/res/sw/dircol/>`_                 |
++-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
+| DYNOPT [2]_ |                     |             |                |                             |      |                        |                   |                 |                                                                                                 |
++-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
+| FROST       | [@Hereid2017]       | Matlab,     | BSD 3-Clause   | Analytic                    | ?    | ?                      | ?                 | IPOPT, fmincon  | `FROST Documentation <http://ayonga.github.io/frost-dev/>`_                                     |
+|             |                     | Mathematica |                |                             |      |                        |                   |                 |                                                                                                 |
++-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
+| GESOP [2]_  |                     |             |                |                             |      |                        |                   |                 |                                                                                                 |
++-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
+| GPOPS       | [@PattersonRao2014] | Matlab      | Commercial     | Automatic differentiation   | No   | Pseudospectral         | No                | SNOPT, IPOPT    | `GPOPS Website <http://www.gpops2.com/>`_                                                       |
++-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
[…]
