### 2018-01-10 csu-hmc/opty#62 @moorepants
@tvdbogert I added OTIS and FROST to the table. See the results here:

https://github.com/csu-hmc/opty/blob/master/paper.md

I moved the table to `paper-table.rst` and use `rst2html` to create the html that I then copy into paper.md, which renders correctly.

There are still gaps in the table, in particular for DYNOPT and GESOP. Can you have a look and see if you know any other particulars. If not, should we leave the caveat that I added about missing info or should we remove them from the table?
