### 2018-01-10 20a0cec
Adde note about which initial guess is used for pen par id example. Fixes #57.
