### 2018-01-10 0ad0c4f
Added example console output for pendulum swind betts examples.
