+++ docs/examples.rst
@@ -15,11 +15,143 @@ Single Pendulum Swing Up
      Your browser does not support the video tag.
    </video>
 
+Example console output::
+
+   ******************************************************************************
+   This program contains Ipopt, a library for large-scale nonlinear optimization.
+    Ipopt is released as open source code under the Eclipse Public License (EPL).
+            For more information visit http://projects.coin-or.org/Ipopt
+   ******************************************************************************
+
+   This is Ipopt version 3.12.8, running with linear solver mumps.
+   NOTE: Other linear solvers might be more efficient (see Ipopt documentation).
+
+   Number of nonzeros in equality constraint Jacobian...:     4994
+   Number of nonzeros in inequality constraint Jacobian.:        0
+   Number of nonzeros in Lagrangian Hessian.............:        0
+
+   Total number of variables............................:     1500
+                        variables with only lower bounds:        0
+                   variables with lower and upper bounds:      500
+                        variables with only upper bounds:        0
+   Total number of equality constraints.................:     1002
+   Total number of inequality constraints...............:        0
+           inequality constraints with only lower bounds:        0
+      inequality constraints with lower and upper bounds:        0
+           inequality constraints with only upper bounds:        0
+
+   iter    objective    inf_pr   inf_du lg(mu)  ||d||  lg(rg) alpha_du alpha_pr  ls
+      0  1.0051250e+01 2.37e+02 8.84e-02   0.0 0.00e+00    -  0.00e+00 0.00e+00   0
+      1  3.9969186e+00 1.83e+02 2.29e+01   0.9 8.83e+00    -  6.88e-01 2.29e-01f  1
+      2  8.3301435e+00 1.38e+02 4.99e+01  -1.2 8.40e+00    -  3.00e-01 2.53e-01h  1
+      3  1.8540696e+01 6.01e+01 5.63e+01  -5.3 6.38e+00    -  4.07e-01 5.70e-01h  1
+      4  1.4859148e+01 3.86e+01 1.07e+02   0.7 5.98e+00    -  2.46e-01 3.63e-01f  1
+      5  1.2535715e+01 2.48e+01 1.26e+02   0.7 4.54e+00    -  5.28e-01 3.68e-01f  1
+      6  1.0337282e+01 2.49e+01 1.53e+02   2.6 1.38e+02    -  5.69e-02 1.96e-02f  1
+      7  1.0132779e+01 1.18e+01 5.70e+02   1.4 4.35e+00    -  7.49e-01 5.39e-01h  1
+      8  1.0049523e+01 7.13e+00 5.63e+02   1.3 5.26e+00    -  5.52e-01 3.95e-01h  1
+      9  1.2555246e+01 4.36e+00 3.29e+02   1.3 4.84e+00    -  6.03e-01 3.96e-01h  1
+   ...
+   iter    objective    inf_pr   inf_du lg(mu)  ||d||  lg(rg) alpha_du alpha_pr  ls
+     80  1.5656141e+01 2.43e-10 2.75e-06 -11.0 4.95e-04    -  1.00e+00 1.00e+00H  1
+     81  1.5656141e+01 2.97e-11 1.01e-05 -11.0 3.08e-04    -  1.00e+00 1.00e+00H  1
+     82  1.5656141e+01 3.06e-11 2.41e-07 -11.0 2.55e-04    -  1.00e+00 1.00e+00H  1
+     83  1.5656141e+01 6.22e-14 3.29e-07 -11.0 1.37e-05    -  1.00e+00 1.00e+00H  1
[…]
