### 2018-01-10 2838dd1
Added references for GESOP (Gath2001) and DYNOPT
