### 2018-01-10 d9fcc7c
Entered DYNOPT and GESOP information in the table
