### 2018-01-10 csu-hmc/opty#62 @tvdbogert
I found some information about DYNOPT and GESOP, I added this with references in the table but I did that in paper.md which breaks your pipeline.  Sorry about that, you may have to copy that back into the .rst file.

I was not sure how you distinguish between DAE and Implicit Dynamics.  Aren't they the same?  Those columns are the same except for DIDO and PSOPT.   DYNOPT allows a mass matrix which is a subclass of DAE/Implicit dynamics.
