### 2018-01-10 5dc6886
Added Betts reference to theory docs.
