### 2018-01-10 csu-hmc/opty#62 @moorepants
Yeah, I wanted to ask you about the DAE thing. In fact, I'm not completely sure that opty can handle DAEs or not. DAEs are differential equations with additional algebraic equations (like a kinematic constraint or something). I'm not sure I have all of the terminogoly precisely correct.

Here is how I have been defining things (which may not be correct);

Implicit First Order Differential equation: f(y'(t), y(t), t) = 0

Explicit First Order Differential equation: y'(t) = f(y(t), t)

Implicit First Order Differential Algebraic Equation: f(y'(t), y(t), t) = 0 and g(y(t), t) = 0

Explicit First Order Differential Algebraic Equation: y'(t) = f(y(t), t) and g(y(t), t) = 0

So I have implicit/explicit forms of both DEs and DAEs.

What should it be?
