### 2018-01-10 csu-hmc/opty#62 @moorepants
There is also this form, which I forget what I've called it:

M y'(t) = f(y(t), t)

The "mass matrix form"?
