+++ paper-table.rst
@@ -10,12 +10,15 @@
 +-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
 | DIRCOL      | [@vonStryk1993]     | Fortran     | Non-commercial | Finite differences          | Yes  | Piecewise linear/cubic | Yes               | NPSOL, SNOPT    | `DIRCOL Website <http://www.sim.informatik.tu-darmstadt.de/en/res/sw/dircol/>`_                 |
 +-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
-| DYNOPT [2]_ |                     |             |                |                             |      |                        |                   |                 |                                                                                                 |
+| DYNOPT      | [@Cizniar2005]      | Matlab      | Custom Open    | Must be supplied by user    | Mass | Pseudospectral         | Mass matrix       | fmincon         | `DYNOPT Code and Documentation <https://bitbucket.org/dynopt/>`_                                |
+|             |                     |             | Source,        |                             | Mat  |                        |                   |                 |                                                                                                 |
+|             |                     |             | Non-commercial |                             |      |                        |                   |                 |                                                                                                 |
 +-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
 | FROST       | [@Hereid2017]       | Matlab,     | BSD 3-Clause   | Analytic                    | ?    | ?                      | ?                 | IPOPT, fmincon  | `FROST Documentation <http://ayonga.github.io/frost-dev/>`_                                     |
 |             |                     | Mathematica |                |                             |      |                        |                   |                 |                                                                                                 |
 +-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
-| GESOP [2]_  |                     |             |                |                             |      |                        |                   |                 |                                                                                                 |
+| GESOP       | [@Gath2001]         | Matlab, C,  | Commercial     | ?                           | No   | Pseudospectral         | No                | SLLSQP, SNOPT,  | `Astos Solutions Gmbh <https://www.astos.de/products/gesop>`_                                   |
+|             |                     | Fortan, Ada |                |                             |      |                        |                   | SOCS            |                                                                                                 |
 +-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
 | GPOPS       | [@PattersonRao2014] | Matlab      | Commercial     | Automatic differentiation   | No   | Pseudospectral         | No                | SNOPT, IPOPT    | `GPOPS Website <http://www.gpops2.com/>`_                                                       |
 +-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
@@ -32,6 +35,4 @@
 | SOCS        | [@Betts2010]        | Fortran     | Commercial     | Finite differences          | Yes  | Euler, RK, & others    | Yes               | built-in        | `SOCS Documentation <http://www.boeing.com/assets/pdf/phantom/socs/docs/SOCS_Users_Guide.pdf>`_ |
 +-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
 
-.. [1] Casadi does not have a built in direct collocation transcription but
-   includes examples which show how to do so for specific problems.
-.. [2] Not enough public information available but included for completeness.
+.. [1] Casadi does not have a built in direct collocation transcription but includes examples which show how to do so for specific problems.
