+++ paper.bib
@@ -291,16 +291,16 @@ Key          = {}
 }
 
 @inproceedings{Cizniar2005,
-  author	 = 	{{\v{C}}i\v{z}niar, M. and Salhi, D. and Fikar, M. and Latifi, M. A.},
-  title	 = 	{DYNOPT - Dynamic Optimisation Code for MATLAB},
-  booktitle	 = 	{13th Annual Conference Proceedings of Technical Computing Prague 2005},
-  year	 = 	{2005},
-  pages	 = 	{},
-  address	 = 	{Praha},
-  publisher	 = 	{Humusoft},
-  annote	 = 	{Dynamic optimisation consists in the determination, in open loop control, of a set of decision variable time profiles (pressure, temperature, flow rate, current, heat duty,~\ldots) for a dynamic system that optimise a given performance index (or cost functional or optimisation criterion)(cost, time, energy, selectivity,~\ldots) subject to specified constraints (safety, environmental and operating constraints). Optimal control refers to the determination of the best time-varying profiles in closed loop control. The orthogonal collocations on finite elements has been developed and implemented within MATLAB environment. The original dynamic optimisation problem is then converted into an NLP problem which is solved using an SQP solver. The gradients of the performance index as well as of the constraints with respect to the parameters are computed numerically and analytically. Six different examples from the literature are considered and treated. The results confirm the optimal values of the literature and the expected performances of the optimisation method when the gradients are computed analytically.},
-  url	 = 	{http://www.kirp.chtf.stuba.sk/publication_info.php?id_pub=255}
-} 
+  author =   {{\v{C}}i\v{z}niar, M. and Salhi, D. and Fikar, M. and Latifi, M. A.},
+  title =   {DYNOPT - Dynamic Optimisation Code for MATLAB},
+  booktitle   =   {13th Annual Conference Proceedings of Technical Computing Prague 2005},
+  year =   {2005},
+  pages =   {},
+  address =   {Praha},
+  publisher =   {Humusoft},
+  annote =   {Dynamic optimisation consists in the determination, in open loop control, of a set of decision variable time profiles (pressure, temperature, flow rate, current, heat duty,~\ldots) for a dynamic system that optimise a given performance index (or cost functional or optimisation criterion)(cost, time, energy, selectivity,~\ldots) subject to specified constraints (safety, environmental and operating constraints). Optimal control refers to the determination of the best time-varying profiles in closed loop control. The orthogonal collocations on finite elements has been developed and implemented within MATLAB environment. The original dynamic optimisation problem is then converted into an NLP problem which is solved using an SQP solver. The gradients of the performance index as well as of the constraints with respect to the parameters are computed numerically and analytically. Six different examples from the literature are considered and treated. The results confirm the optimal values of the literature and the expected performances of the optimisation method when the gradients are computed analytically.},
+  url   =   {http://www.kirp.chtf.stuba.sk/publication_info.php?id_pub=255}
+}
 
 @inproceedings{Gath2001,
   author = {Gath, Peter and Well, Klaus},
