### 2018-01-10 f4311a1
Added parallel feature to README.
