### 2018-01-10 csu-hmc/opty#62 @tvdbogert
Yes, that's the mass matrix form.  If M is non-invertible you effectively have a DAE.

The implicit form is a DAE if the mass matrix df/dxdot is non-invertible, otherwise it's an implicit ODE.

Opty can handle DAE (and I have done that, but not with opty) as long as one of these two conditions is true: (1) the Jacobian is such that the algebraic variables could be solved from the differential variables, or (2) the cost function ensures that the algebraic variables are uniquely solved.  The latter is basically the same reason why the unknown control variables being in the cost function.

Maybe, for simplicity, we should not talk about DAE capability  (then we have to get into index 1, 2, etc and be careful with terminology).  I suggest we only say whether implicit formulation is supported.  That's a simply a formulation and API issue, nothing to do with invertability.  etc.  Then we can say "yes" or "mass-matrix form" or "no" in the "implicit formulation" column.

Even an index-3 DAE (i.e. a multibody system with additional kinematic constraints) could be solved by opty if the cost function includes all of the (algebraic) reaction force variables.  This is another confirmation that DAE capability has nothing to do with the software or algorithm.  It also depends on the cost function.
