### 2018-01-10 d199cc0
Inverted pendulum now works with PyDy 0.3+ and SymPy 1.0+ and Python 2 and 3. Fixes #68, Fixes #54.
