### 2018-01-10 csu-hmc/opty#62 @moorepants
That all sounds fine. I will remove the DAE column.

As a side note, my thoughts were that it is worth describing the implicit/explicit form regardless of if you have a DAE or not because removing the need to both form the mass matrix and solve for the state derivatives removes a fairly large computational burden which is advantageous with ODEs and DAEs.
