### 2018-01-10 csu-hmc/opty#62 @tvdbogert
Yes, that would be worth mentioning because opty has this capability and others don't.

Some of these tools have the limitation that they don't scale well to large problems or high-dimensional state spaces.  But there is no way we can make specific statements.  Maybe a general cautionary note.
