+++ paper.bib
@@ -219,6 +219,16 @@ Key          = {}
   urldate    = {2017-06-04},
 }
 
+@article{Hargraves1987,
+  title = {Direct Trajectory Optimization Using Nonlinear Programming and Collocation},
+  journal = {AIAA Journal of Guidance, Control, and Dynamics},
+  volume = {10},
+  number = {4},
+  year = {1987},
+  pages = {338--342},
+  author = {Hargraves, C. R. and Paris, S. W.},
+}
+
 @article{Behnel2011,
   title = {Cython: {The} {Best} of {Both} {Worlds}},
   volume = {13},
