+++ paper-table.rst
@@ -1,38 +1,38 @@
 
-+-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
-| Name        | Citation            | Language    | License        | Derivatives                 | DAEs |  Discretization        | Implicit Dynamics | Solvers         | Project Website                                                                                 |
-+=============+=====================+=============+================+=============================+======+========================+===================+=================+=================================================================================================+
-| Casadi [1]_ | [@Andersson2013]    | C++,        | LGPL           | Automatic differentiation   | Yes  | None                   | Yes               | IPOPT, WORHP,   | `Casadi Website <https://github.com/casadi/casadi/wiki>`_                                       |
-|             |                     | Python,     |                |                             |      |                        |                   | SNOPT, KNITRO   |                                                                                                 |
-|             |                     | Octave,     |                |                             |      |                        |                   |                 |                                                                                                 |
-+-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
-| DIDO        | [@Ross2002]         | Matlab      | Commercial     | Analytic                    | No   | Pseudospectral         | Yes               | built-in        | `DIDO Website <http://www.elissarglobal.com/industry/products/software-3/>`_                    |
-+-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
-| DIRCOL      | [@vonStryk1993]     | Fortran     | Non-commercial | Finite differences          | Yes  | Piecewise linear/cubic | Yes               | NPSOL, SNOPT    | `DIRCOL Website <http://www.sim.informatik.tu-darmstadt.de/en/res/sw/dircol/>`_                 |
-+-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
-| DYNOPT      | [@Cizniar2005]      | Matlab      | Custom Open    | Must be supplied by user    | Mass | Pseudospectral         | Mass matrix       | fmincon         | `DYNOPT Code and Documentation <https://bitbucket.org/dynopt/>`_                                |
-|             |                     |             | Source,        |                             | Mat  |                        |                   |                 |                                                                                                 |
-|             |                     |             | Non-commercial |                             |      |                        |                   |                 |                                                                                                 |
-+-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
-| FROST       | [@Hereid2017]       | Matlab,     | BSD 3-Clause   | Analytic                    | ?    | ?                      | ?                 | IPOPT, fmincon  | `FROST Documentation <http://ayonga.github.io/frost-dev/>`_                                     |
-|             |                     | Mathematica |                |                             |      |                        |                   |                 |                                                                                                 |
-+-------------+---------------------+-------------+----------------+-----------------------------+------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
-| GESOP       | [@Gath2001]         | Matlab, C,  | Commercial     | ?                           | No   | Pseudospectral         | No                | SLLSQP, SNOPT,  | `Astos Solutions Gmbh <https://www.astos.de/products/gesop>`_                                   |
-|             |                     | Fortan, Ada |                |                             |      |                        |                   | SOCS            |                                                                                                 |
[…]
