### 2018-01-10 csu-hmc/opty#62 @moorepants
Ok, the DAE column is gone and I added an OTIS citation. I'm going to close this.
