### 2018-01-10 205c2ea
Turn back on animation for pend swing example.
