### 2018-01-10 #300 @moorepants
We have made all of the changes requested by the reviewers. You can see an issue for each reviewer comment here:

https://github.com/csu-hmc/opty/issues?q=is%3Aissue+is%3Aclosed+label%3Ajoss

And pull requests that close all of these issues here:

https://github.com/csu-hmc/opty/pulls?q=is%3Apr+is%3Aclosed+label%3Ajoss

I have not made a new release of opty yet, so testing from the master branch is required. I would prefer to make the release after the review is done so that I can incorporate all of the changes into the release.

The documentation reflects the master branch:

https://github.com/csu-hmc/opty/pulls?q=is%3Apr+is%3Aclosed+label%3Ajoss

Thanks for your patience.

cc @tvdbogert
