### 2018-01-10 2d087a6
Fixed ipopt citation.
