### 2018-01-10 a2af3c3
Fixed period in bib.
