+++ paper.bib
@@ -314,7 +314,7 @@ Key          = {}
 
 @inproceedings{Gath2001,
   author = {Gath, Peter and Well, Klaus},
-  title = {Trajectory optimization using a combination of direct multiple shooting and collocation}.
+  title = {Trajectory optimization using a combination of direct multiple shooting and collocation},
   year = {2001},
   doi = {10.2514/6.2001-4047},
   url = {https://doi.org/10.2514/6.2001-4047},
