### 2018-01-10 #300 @moorepants
Editors, I switched to an HTML table in the markdown file since HTML can be put into markdown files but the PDF does not render the table correctly. Any suggestions on getting a table in the paper?
