### 2018-01-11 28258ad
One more pass through the paper.
