+++ paper.md
@@ -27,23 +27,23 @@ bibliography: paper.bib
 
 opty is a tool for describing and solving trajectory optimization and parameter
 identification problems based on symbolic descriptions of ordinary differential
-equations that describe a dynamical system. The motivation for its development
-resides in the need to solve optimal control problems of biomechanical systems.
-The target audience is engineers and scientists interested in solving nonlinear
-optimal control and parameter identification problems with minimal
-computational overhead.
+equations and differential algebriac equations that describe a dynamical
+system. The motivation for its development resides in the need to solve optimal
+control problems of biomechanical systems. The target audience is engineers
+and scientists interested in solving nonlinear optimal control and parameter
+identification problems with minimal computational overhead.
 
-A user of opty is responsible for specifying the system's dynamics (the ODEs),
-the cost or fitness function, bounds on the solution, and the initial guess for
-the solution. opty uses this problem specification to derive the constraints
-needed to solve the optimization problem using the direct collocation method
-[@Betts2010]. This method maps the problem to a non-linear programming problem
-and the result is then solved numerically with an interior point optimizer,
-IPOPT [@Wachter2006] which is wrapped by cyipopt [@Cyipopt2017] for use in
-Python. The software allows the user to describe the dynamical system of
-interest at a high level in symbolic form without needing to concern themselves
-with the numerical computation details. This is made possible by utilizing
-SymPy [@Meurer2017] and Cython [@Behnel2011] for code generation and
+A user of opty is responsible for specifying the system's dynamics (the
+ODE/DAEs), the cost or fitness function, bounds on the solution, and the
+initial guess for the solution. opty uses this problem specification to derive
+the constraints needed to solve the optimization problem using the direct
+collocation method [@Betts2010]. This method maps the problem to a non-linear
+programming problem and the result is then solved numerically with an interior
+point optimizer IPOPT [@Wachter2006] which is wrapped by cyipopt [@Cyipopt2017]
+for use in Python. The software allows the user to describe the dynamical
+system of interest at a high level in symbolic form without needing to concern
+themselves with the numerical computation details. This is made possible by
+utilizing SymPy [@Meurer2017] and Cython [@Behnel2011] for code generation and
 just-in-time compilation to obtain wrap optimized C functions that are
 accessible in Python.
 
@@ -73,7 +73,8 @@ globally optimal trajectory in a nonlinear system when a good initial guess is
 not available [@Zarei2016].
 
 There are existing software packages that can solve optimal control problems
-and have have similarities to opty. Below, is a feature comparison:
+and have have similarities to opty. Below, is a feature comparison of those we
+are aware of:
 
 <table border="1" class="docutils">
 <colgroup>
@@ -242,17 +243,18 @@ Sparse finite differences</td>
 
 Each of these software packages offer a different combination of attributes and
 features that make it useful for different problems. opty is the only package
-that has a liberal open source license, following precedent set by other core
-scientific Python packages. This allows anyone to use and modify the code
-without having to share the source of their application. opty also is the only
-package, open source or not, that allows (in fact forces) the user to describe
-their problem via a high level symbolic mathematical description using the API
-of a widely used computer algebra system. This relieves the user from having to
+that has a liberal open source license for itself and its dependencies,
+following precedent set by other core scientific Python packages. This allows
+anyone to use and modify the code without having to share the source of their
+application. opty also is the only package, open source or not, that allows (in
+fact forces) the user to describe their problem via a high level symbolic
+mathematical description using the API of a widely used computer algebra system
+instead of a domain specific language.  This relieves the user from having to
 translate the much simpler continuous problem definition into a discretize NLP
[…]
