### 2018-01-11 41c09a5
Removed dangling phrase.
