+++ paper.md
@@ -249,15 +249,15 @@ anyone to use and modify the code without having to share the source of their
 application. opty also is the only package, open source or not, that allows (in
 fact forces) the user to describe their problem via a high level symbolic
 mathematical description using the API of a widely used computer algebra system
-instead of a domain specific language.  This relieves the user from having to
+instead of a domain specific language. This relieves the user from having to
 translate the much simpler continuous problem definition into a discretize NLP
 problem. opty leverages the popular Scientific Python core tools like NumPy,
 SymPy, Cython, and matplotlib allowing users to include opty code into Python
-programs.  set of tools. Lastly, the numeric code generated by opty to evaluate
-the NLP constraints is optimized providing extremely efficient parallel
-evaluation of the contraints. This becomes very valuable for high dimensional
-dynamics. opty currently does not offer a wide range of discretization methods
-nor support for solvers other than IPOPT, but those could relatively easily be
-added based on user need.
+programs. Lastly, the numeric code generated by opty to evaluate the NLP
+constraints is optimized providing extremely efficient parallel evaluation of
+the contraints. This becomes very valuable for high dimensional dynamics. opty
+currently does not offer a wide range of discretization methods nor support for
+solvers other than IPOPT, but those could relatively easily be added based on
+user need.
 
 # References
