### 2018-01-16 #300 @Kevin-Mattheus-Moerman
@moorepants @arfon I tested the local paper build myself.  I was able to get it to create markdown based tables without any issues. What was your issue? Do you perhaps have a table syntax error somewhere?

I did notice that your particular table will spill beyond the margins using pandoc (for both HTML and markdown).  [Here is a link](https://pandoc.org/MANUAL.html#pandocs-markdown) to how to create tables in pandoc markdown. HTML is not listed there and I could not get that to work.

Is it an idea to render a picture of the table in the right layout and to render the image in markdown instead instead? I suppose the references would not show up.... sigh... this is difficult!
