### 2018-01-16 #300 @arfon
I wonder if we could put the table on its own page and hack the LaTeX template to make this happen. Some ideas here: https://stackoverflow.com/questions/25849814/rstudio-rmarkdown-both-portrait-and-landscape-layout-in-a-single-pdf
