### 2018-01-17 #300 @karthik
@arfon I don't have a good solution unfortunately. :(
