### 2018-01-17 #300 @arfon
> Is it possible to use the markdown "grid-tables" extension? I already have the table typed up like that. If not I can try converting the first table option on the pandoc markdown documentation.

Sure thing. Let's give that a go. Are those commits in `master`?
