### 2018-01-24 71c124b
Now using footnotes and grid-tables extensions in the paper.md file.
