+++ paper.md
@@ -76,170 +76,43 @@ There are existing software packages that can solve optimal control problems
 and have have similarities to opty. Below, is a feature comparison of those we
 are aware of:
 
-<table border="1" class="docutils">
-<colgroup>
-<col width="5%" />
-<col width="8%" />
-<col width="5%" />
-<col width="6%" />
-<col width="12%" />
-<col width="10%" />
-<col width="8%" />
-<col width="7%" />
-<col width="39%" />
-</colgroup>
-<thead valign="bottom">
-<tr><th class="head">Name</th>
-<th class="head">Citation</th>
-<th class="head">Language</th>
-<th class="head">License</th>
-<th class="head">Derivatives</th>
-<th class="head">Discretization</th>
-<th class="head">Implicit Dynamics</th>
-<th class="head">Solvers</th>
-<th class="head">Project Website</th>
-</tr>
-</thead>
-<tbody valign="top">
-<tr><td>Casadi <a class="footnote-reference" href="#id2" id="id1">[1]</a></td>
-<td>[&#64;Andersson2013]</td>
-<td>C++,
-Python,
-Octave,</td>
-<td>LGPL</td>
-<td>Automatic differentiation</td>
-<td>None</td>
-<td>Yes</td>
-<td>IPOPT, WORHP,
-SNOPT, KNITRO</td>
-<td><a class="reference external" href="https://github.com/casadi/casadi/wiki">Casadi Website</a></td>
-</tr>
-<tr><td>DIDO</td>
-<td>[&#64;Ross2002]</td>
-<td>Matlab</td>
-<td>Commercial</td>
-<td>Analytic</td>
-<td>Pseudospectral</td>
-<td>Yes</td>
-<td>built-in</td>
-<td><a class="reference external" href="http://www.elissarglobal.com/industry/products/software-3/">DIDO Website</a></td>
-</tr>
-<tr><td>DIRCOL</td>
-<td>[&#64;vonStryk1993]</td>
-<td>Fortran</td>
-<td>Non-commercial</td>
-<td>Finite differences</td>
-<td>Piecewise linear/cubic</td>
-<td>Yes</td>
-<td>NPSOL, SNOPT</td>
-<td><a class="reference external" href="http://www.sim.informatik.tu-darmstadt.de/en/res/sw/dircol/">DIRCOL Website</a></td>
-</tr>
-<tr><td>DYNOPT</td>
-<td>[&#64;Cizniar2005]</td>
-<td>Matlab</td>
-<td>Custom Open
-Source,
-Non-commercial</td>
-<td>Must be supplied by user</td>
-<td>Pseudospectral</td>
-<td>Mass matrix</td>
-<td>fmincon</td>
-<td><a class="reference external" href="https://bitbucket.org/dynopt/">DYNOPT Code and Documentation</a></td>
-</tr>
-<tr><td>FROST</td>
-<td>[&#64;Hereid2017]</td>
-<td>Matlab,
-Mathematica</td>
-<td>BSD 3-Clause</td>
-<td>Analytic</td>
-<td>?</td>
-<td>?</td>
-<td>IPOPT, fmincon</td>
-<td><a class="reference external" href="http://ayonga.github.io/frost-dev/">FROST Documentation</a></td>
-</tr>
-<tr><td>GESOP</td>
-<td>[&#64;Gath2001]</td>
-<td>Matlab, C,
-Fortan, Ada</td>
-<td>Commercial</td>
-<td>?</td>
-<td>Pseudospectral</td>
-<td>No</td>
-<td>SLLSQP, SNOPT,
[…]
