### 2018-01-24 #300 @moorepants
I have updated the `paper.md` file to make use of the footnotes and grid-tables extensions listed here:

https://pandoc.org/MANUAL.html
