### 2018-01-24 #300 @moorepants
It is close now. Is it possible to choose a smaller font for the table? This is possible in LaTeX.
