+++ paper.md
@@ -85,7 +85,7 @@ are aware of:
 +-------------+---------------------+-------------+----------------+-----------------------------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
 | DIDO        | [@Ross2002]         | Matlab      | Commercial     | Analytic                    | Pseudospectral         | Yes               | built-in        | [DIDO Website](http://www.elissarglobal.com/industry/products/software-3/)                      |
 +-------------+---------------------+-------------+----------------+-----------------------------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
-| DIRCOL      | [@vonStryk1993]     | Fortran     | Non-commercial | Finite differences          | Piecewise linear/cubic | Yes               | NPSOL, SNOPT    | [DIRCOL Website](http://www.sim.informatik.tu-darmstadt.de/en/res/sw/dircol/)                   |
+| DIRCOL      | [@vonStryk1992]     | Fortran     | Non-commercial | Finite differences          | Piecewise linear/cubic | Yes               | NPSOL, SNOPT    | [DIRCOL Website](http://www.sim.informatik.tu-darmstadt.de/en/res/sw/dircol/)                   |
 +-------------+---------------------+-------------+----------------+-----------------------------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
 | DYNOPT      | [@Cizniar2005]      | Matlab      | Custom Open    | Must be supplied by user    | Pseudospectral         | Mass matrix       | fmincon         | [DYNOPT Code and Documentation](https://bitbucket.org/dynopt/)                                  |
 |             |                     |             | Source,        |                             |                        |                   |                 |                                                                                                 |
