### 2018-01-24 #300 @moorepants
I just found this: https://pandoc.org/MANUAL.html#extension-raw_tex

It may be easier to just write the document in LaTeX using this extension. I'll give it a try too. I haven't been able to figure out how to manipulate the font size of the table using markdown.
