### 2018-01-29 1dd5b7b
Trying raw latex commands in the markdown file.
