+++ paper.md
@@ -89,7 +89,7 @@ are aware of:
 +-------------+---------------------+-------------+----------------+-----------------------------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
 | DIRCOL      | [@vonStryk1992]     | Fortran     | Non-commercial | Finite differences          | Piecewise linear/cubic | Yes               | NPSOL, SNOPT    | [DIRCOL Website](http://www.sim.informatik.tu-darmstadt.de/en/res/sw/dircol/)                   |
 +-------------+---------------------+-------------+----------------+-----------------------------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
-| DYNOPT      | [@Cizniar2005]      | Matlab      | Custom Open    | Must be supplied by user    | Pseudospectral         | Mass matrix       | fmincon         | [DYNOPT Code and Documentation](https://bitbucket.org/dynopt/)                                  |
+| DYNOPT      | [@Cizniar2005]      | Matlab      | Custom Open    | Must be supplied by user    | Pseudospectral         | Mass matrix       | fmincon         | [DYNOPT Code](https://bitbucket.org/dynopt/)                                                    |
 |             |                     |             | Source,        |                             |                        |                   |                 |                                                                                                 |
 |             |                     |             | Non-commercial |                             |                        |                   |                 |                                                                                                 |
 +-------------+---------------------+-------------+----------------+-----------------------------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
