### 2018-01-29 1f4523f
Moved link targets outside of table to try to make last column more narrow.
