+++ paper.md
@@ -78,43 +78,56 @@ are aware of:
 
 \tiny
 
-+-------------+---------------------+-------------+----------------+-----------------------------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
-| Name        | Citation            | Language    | License        | Derivatives                 |  Discretization        | Implicit Dynamics | Solvers         | Project Website                                                                                 |
-+=============+=====================+=============+================+=============================+========================+===================+=================+=================================================================================================+
-| Casadi [^1] | [@Andersson2013]    | C++,        | LGPL           | Automatic differentiation   | None                   | Yes               | IPOPT, WORHP,   | [Casadi Website](https://github.com/casadi/casadi/wiki)                                         |
-|             |                     | Python,     |                |                             |                        |                   | SNOPT, KNITRO   |                                                                                                 |
-|             |                     | Octave,     |                |                             |                        |                   |                 |                                                                                                 |
-+-------------+---------------------+-------------+----------------+-----------------------------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
-| DIDO        | [@Ross2002]         | Matlab      | Commercial     | Analytic                    | Pseudospectral         | Yes               | built-in        | [DIDO Website](http://www.elissarglobal.com/industry/products/software-3/)                      |
-+-------------+---------------------+-------------+----------------+-----------------------------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
-| DIRCOL      | [@vonStryk1992]     | Fortran     | Non-commercial | Finite differences          | Piecewise linear/cubic | Yes               | NPSOL, SNOPT    | [DIRCOL Website](http://www.sim.informatik.tu-darmstadt.de/en/res/sw/dircol/)                   |
-+-------------+---------------------+-------------+----------------+-----------------------------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
-| DYNOPT      | [@Cizniar2005]      | Matlab      | Custom Open    | Must be supplied by user    | Pseudospectral         | Mass matrix       | fmincon         | [DYNOPT Code](https://bitbucket.org/dynopt/)                                                    |
-|             |                     |             | Source,        |                             |                        |                   |                 |                                                                                                 |
-|             |                     |             | Non-commercial |                             |                        |                   |                 |                                                                                                 |
-+-------------+---------------------+-------------+----------------+-----------------------------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
-| FROST       | [@Hereid2017]       | Matlab,     | BSD 3-Clause   | Analytic                    | ?                      | ?                 | IPOPT, fmincon  | [FROST Documentation](http://ayonga.github.io/frost-dev/)                                       |
-|             |                     | Mathematica |                |                             |                        |                   |                 |                                                                                                 |
-+-------------+---------------------+-------------+----------------+-----------------------------+------------------------+-------------------+-----------------+-------------------------------------------------------------------------------------------------+
-| GESOP       | [@Gath2001]         | Matlab, C,  | Commercial     | ?                           | Pseudospectral         | No                | SLLSQP, SNOPT,  | [Astos Solutions Gmbh](https://www.astos.de/products/gesop)                                     |
[…]
