### 2018-01-30 #300 @moorepants
The table fits now. It isn't perfect but it is too difficult to try to fine tune it via markdown and limited control of the LaTeX source. We can go with this.
