### 2018-01-30 #300 @arfon
OK thanks @moorepants. @Kevin-Mattheus-Moerman - is this submission good to accept now? The checklist at the top of the review isn't filled but it sounds like we might be done based on other comments in the thread.
