### 2018-01-30 #300 @stavness
Looks great!. I went through the checklist and checked all the boxes.

On Tue, Jan 30, 2018 at 5:35 AM, Arfon Smith <notifications@github.com>
wrote:

> OK thanks @moorepants <https://github.com/moorepants>.
> @Kevin-Mattheus-Moerman <https://github.com/kevin-mattheus-moerman> - is
> this submission good to accept now? The checklist at the top of the review
> isn't filled but it sounds like we might be done based on other comments in
> the thread.
>
