### 2018-01-30 #300 @moorepants
I will package a new release of the software, push to PyPi, conda, and then Zenodo for archival purposes. Note that the version number will go up.

Do we not need to get @jslee02 @cybanical and @demotu's final approval? I addressed comments from all of them in the new version.
