### 2018-01-30 09042a9
Removed the tex file.
