### 2018-01-30 5277772
Fixed changelog version numbers and added notes for version 1.0.0.
