+++ CHANGELOG.rst
@@ -1,23 +1,28 @@
+Version 1.0.0
+=============
+
+- Added JOSS paper.
+- Added theory section to the documentation.
+- Added optional parallel execution if openmp is installed.
+- Fixed a bug in plot_trajectories.
+- Adjusted the pendulum swing up torque bounds.
+- Updated examples to work with newer dependency versions.
+
 Version 0.2.0
 =============
 
 - Added Sphinx documentation and Read The Docs integration.
 - Added plotting to the Problem class and matplotlib as an optional dependency.
 - Added conda forge installation instructions and Anaconda badge to the README.
-- Bumped PyDy dep to 0.3.0 and SymPy dep to 1.0.0.
-
-Version 1.2.0
-=============
-
 - park2004 example now works with both Python 2.7 and 3.5+.
 - Bumped the min dependencies for SymPy and PyDy to 1.0.0 and 0.3.0.
 
-Version 1.1.1
+Version 0.1.1
 =============
 
-Added a MANIFEST.in file.
+- Added a MANIFEST.in file.
 
 Version 0.1.0
 =============
 
-Initial release.
+- Initial release.
