### 2018-01-30 9789b3f
Bumped version to 1.0.0.
