+++ opty/version.py
@@ -1,3 +1,3 @@
 #!/usr/bin/env python
 
-__version__ = '0.3.0.dev0'
+__version__ = '1.0.0'
