### 2018-01-30 #300 @moorepants
I've pushed version 1.0.0 to Zenodo:

DOI: 10.5281/zenodo.1162870

https://zenodo.org/record/1162870
