### 2018-01-30 8514e70
Added a Zenodo badge for version 1.0.0.
